@@ -1,6 +1,6 @@
 module github.com/cshep4/premier-predictor-microservices/src/common
 
-go 1.13
+go 1.14
 
 replace github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen => ../../proto-gen/model/gen
 
@@ -8,8 +8,8 @@ require (
 	cloud.google.com/go v0.53.0
 	contrib.go.opencensus.io/exporter/stackdriver v0.13.0
 	github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen v0.0.0-00010101000000-000000000000
-	github.com/golang/mock v1.4.0
-	github.com/golang/protobuf v1.3.3
+	github.com/golang/mock v1.4.4
+	github.com/golang/protobuf v1.4.1
 	github.com/gorilla/handlers v1.4.2
 	github.com/gorilla/mux v1.7.3
 	github.com/grpc-ecosystem/go-grpc-middleware v1.2.0
