@@ -1,29 +1,35 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.25.0
+// 	protoc        v3.12.3
 // source: chat.proto
 
 package model
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
-import empty "github.com/golang/protobuf/ptypes/empty"
-import timestamp "github.com/golang/protobuf/ptypes/timestamp"
-
 import (
-	context "golang.org/x/net/context"
+	context "context"
+	proto "github.com/golang/protobuf/proto"
+	empty "github.com/golang/protobuf/ptypes/empty"
+	timestamp "github.com/golang/protobuf/ptypes/timestamp"
 	grpc "google.golang.org/grpc"
+	codes "google.golang.org/grpc/codes"
+	status "google.golang.org/grpc/status"
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
 )
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+// This is a compile-time assertion that a sufficiently up-to-date version
+// of the legacy proto package is being used.
+const _ = proto.ProtoPackageIsVersion4
 
 type Message_MsgType int32
 
@@ -32,402 +38,739 @@ const (
 	Message_MESSAGE Message_MsgType = 1
 )
 
-var Message_MsgType_name = map[int32]string{
-	0: "READ",
-	1: "MESSAGE",
-}
-var Message_MsgType_value = map[string]int32{
-	"READ":    0,
-	"MESSAGE": 1,
+// Enum value maps for Message_MsgType.
+var (
+	Message_MsgType_name = map[int32]string{
+		0: "READ",
+		1: "MESSAGE",
+	}
+	Message_MsgType_value = map[string]int32{
+		"READ":    0,
+		"MESSAGE": 1,
+	}
+)
+
+func (x Message_MsgType) Enum() *Message_MsgType {
+	p := new(Message_MsgType)
+	*p = x
+	return p
 }
 
 func (x Message_MsgType) String() string {
-	return proto.EnumName(Message_MsgType_name, int32(x))
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
 }
+
+func (Message_MsgType) Descriptor() protoreflect.EnumDescriptor {
+	return file_chat_proto_enumTypes[0].Descriptor()
+}
+
+func (Message_MsgType) Type() protoreflect.EnumType {
+	return &file_chat_proto_enumTypes[0]
+}
+
+func (x Message_MsgType) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use Message_MsgType.Descriptor instead.
 func (Message_MsgType) EnumDescriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{6, 0}
+	return file_chat_proto_rawDescGZIP(), []int{6, 0}
 }
 
 type AddRequest struct {
-	UserId               string   `protobuf:"bytes,1,opt,name=userId,proto3" json:"userId,omitempty"`
-	ChatId               string   `protobuf:"bytes,2,opt,name=chatId,proto3" json:"chatId,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	UserId string `protobuf:"bytes,1,opt,name=userId,proto3" json:"userId,omitempty"`
+	ChatId string `protobuf:"bytes,2,opt,name=chatId,proto3" json:"chatId,omitempty"`
 }
 
-func (m *AddRequest) Reset()         { *m = AddRequest{} }
-func (m *AddRequest) String() string { return proto.CompactTextString(m) }
-func (*AddRequest) ProtoMessage()    {}
+func (x *AddRequest) Reset() {
+	*x = AddRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[0]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *AddRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*AddRequest) ProtoMessage() {}
+
+func (x *AddRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[0]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use AddRequest.ProtoReflect.Descriptor instead.
 func (*AddRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{0}
-}
-func (m *AddRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_AddRequest.Unmarshal(m, b)
-}
-func (m *AddRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_AddRequest.Marshal(b, m, deterministic)
-}
-func (dst *AddRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_AddRequest.Merge(dst, src)
-}
-func (m *AddRequest) XXX_Size() int {
-	return xxx_messageInfo_AddRequest.Size(m)
-}
-func (m *AddRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_AddRequest.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{0}
 }
 
-var xxx_messageInfo_AddRequest proto.InternalMessageInfo
-
-func (m *AddRequest) GetUserId() string {
-	if m != nil {
-		return m.UserId
+func (x *AddRequest) GetUserId() string {
+	if x != nil {
+		return x.UserId
 	}
 	return ""
 }
 
-func (m *AddRequest) GetChatId() string {
-	if m != nil {
-		return m.ChatId
+func (x *AddRequest) GetChatId() string {
+	if x != nil {
+		return x.ChatId
 	}
 	return ""
 }
 
 type LatestMessagesRequest struct {
-	ChatId               string   `protobuf:"bytes,1,opt,name=chatId,proto3" json:"chatId,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	ChatId string `protobuf:"bytes,1,opt,name=chatId,proto3" json:"chatId,omitempty"`
 }
 
-func (m *LatestMessagesRequest) Reset()         { *m = LatestMessagesRequest{} }
-func (m *LatestMessagesRequest) String() string { return proto.CompactTextString(m) }
-func (*LatestMessagesRequest) ProtoMessage()    {}
+func (x *LatestMessagesRequest) Reset() {
+	*x = LatestMessagesRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[1]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *LatestMessagesRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*LatestMessagesRequest) ProtoMessage() {}
+
+func (x *LatestMessagesRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[1]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use LatestMessagesRequest.ProtoReflect.Descriptor instead.
 func (*LatestMessagesRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{1}
-}
-func (m *LatestMessagesRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_LatestMessagesRequest.Unmarshal(m, b)
-}
-func (m *LatestMessagesRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_LatestMessagesRequest.Marshal(b, m, deterministic)
-}
-func (dst *LatestMessagesRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_LatestMessagesRequest.Merge(dst, src)
-}
-func (m *LatestMessagesRequest) XXX_Size() int {
-	return xxx_messageInfo_LatestMessagesRequest.Size(m)
-}
-func (m *LatestMessagesRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_LatestMessagesRequest.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{1}
 }
 
-var xxx_messageInfo_LatestMessagesRequest proto.InternalMessageInfo
-
-func (m *LatestMessagesRequest) GetChatId() string {
-	if m != nil {
-		return m.ChatId
+func (x *LatestMessagesRequest) GetChatId() string {
+	if x != nil {
+		return x.ChatId
 	}
 	return ""
 }
 
 type PreviousMessagesRequest struct {
-	ChatId               string   `protobuf:"bytes,1,opt,name=chatId,proto3" json:"chatId,omitempty"`
-	CurrentMessageId     string   `protobuf:"bytes,2,opt,name=currentMessageId,proto3" json:"currentMessageId,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	ChatId           string `protobuf:"bytes,1,opt,name=chatId,proto3" json:"chatId,omitempty"`
+	CurrentMessageId string `protobuf:"bytes,2,opt,name=currentMessageId,proto3" json:"currentMessageId,omitempty"`
 }
 
-func (m *PreviousMessagesRequest) Reset()         { *m = PreviousMessagesRequest{} }
-func (m *PreviousMessagesRequest) String() string { return proto.CompactTextString(m) }
-func (*PreviousMessagesRequest) ProtoMessage()    {}
+func (x *PreviousMessagesRequest) Reset() {
+	*x = PreviousMessagesRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[2]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *PreviousMessagesRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*PreviousMessagesRequest) ProtoMessage() {}
+
+func (x *PreviousMessagesRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[2]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use PreviousMessagesRequest.ProtoReflect.Descriptor instead.
 func (*PreviousMessagesRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{2}
-}
-func (m *PreviousMessagesRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_PreviousMessagesRequest.Unmarshal(m, b)
-}
-func (m *PreviousMessagesRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_PreviousMessagesRequest.Marshal(b, m, deterministic)
-}
-func (dst *PreviousMessagesRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_PreviousMessagesRequest.Merge(dst, src)
-}
-func (m *PreviousMessagesRequest) XXX_Size() int {
-	return xxx_messageInfo_PreviousMessagesRequest.Size(m)
-}
-func (m *PreviousMessagesRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_PreviousMessagesRequest.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{2}
 }
 
-var xxx_messageInfo_PreviousMessagesRequest proto.InternalMessageInfo
-
-func (m *PreviousMessagesRequest) GetChatId() string {
-	if m != nil {
-		return m.ChatId
+func (x *PreviousMessagesRequest) GetChatId() string {
+	if x != nil {
+		return x.ChatId
 	}
 	return ""
 }
 
-func (m *PreviousMessagesRequest) GetCurrentMessageId() string {
-	if m != nil {
-		return m.CurrentMessageId
+func (x *PreviousMessagesRequest) GetCurrentMessageId() string {
+	if x != nil {
+		return x.CurrentMessageId
 	}
 	return ""
 }
 
 type MessageList struct {
-	Messages             []*Message `protobuf:"bytes,1,rep,name=messages,proto3" json:"messages,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}   `json:"-"`
-	XXX_unrecognized     []byte     `json:"-"`
-	XXX_sizecache        int32      `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Messages []*Message `protobuf:"bytes,1,rep,name=messages,proto3" json:"messages,omitempty"`
 }
 
-func (m *MessageList) Reset()         { *m = MessageList{} }
-func (m *MessageList) String() string { return proto.CompactTextString(m) }
-func (*MessageList) ProtoMessage()    {}
+func (x *MessageList) Reset() {
+	*x = MessageList{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[3]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *MessageList) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*MessageList) ProtoMessage() {}
+
+func (x *MessageList) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[3]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use MessageList.ProtoReflect.Descriptor instead.
 func (*MessageList) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{3}
-}
-func (m *MessageList) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_MessageList.Unmarshal(m, b)
-}
-func (m *MessageList) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_MessageList.Marshal(b, m, deterministic)
-}
-func (dst *MessageList) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_MessageList.Merge(dst, src)
-}
-func (m *MessageList) XXX_Size() int {
-	return xxx_messageInfo_MessageList.Size(m)
-}
-func (m *MessageList) XXX_DiscardUnknown() {
-	xxx_messageInfo_MessageList.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{3}
 }
 
-var xxx_messageInfo_MessageList proto.InternalMessageInfo
-
-func (m *MessageList) GetMessages() []*Message {
-	if m != nil {
-		return m.Messages
+func (x *MessageList) GetMessages() []*Message {
+	if x != nil {
+		return x.Messages
 	}
 	return nil
 }
 
 type SendRequest struct {
-	Message              string               `protobuf:"bytes,1,opt,name=message,proto3" json:"message,omitempty"`
-	UserId               string               `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
-	ChatId               string               `protobuf:"bytes,3,opt,name=chatId,proto3" json:"chatId,omitempty"`
-	DateTime             *timestamp.Timestamp `protobuf:"bytes,4,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
-	XXX_unrecognized     []byte               `json:"-"`
-	XXX_sizecache        int32                `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Message  string               `protobuf:"bytes,1,opt,name=message,proto3" json:"message,omitempty"`
+	UserId   string               `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
+	ChatId   string               `protobuf:"bytes,3,opt,name=chatId,proto3" json:"chatId,omitempty"`
+	DateTime *timestamp.Timestamp `protobuf:"bytes,4,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
 }
 
-func (m *SendRequest) Reset()         { *m = SendRequest{} }
-func (m *SendRequest) String() string { return proto.CompactTextString(m) }
-func (*SendRequest) ProtoMessage()    {}
+func (x *SendRequest) Reset() {
+	*x = SendRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *SendRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*SendRequest) ProtoMessage() {}
+
+func (x *SendRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use SendRequest.ProtoReflect.Descriptor instead.
 func (*SendRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{4}
-}
-func (m *SendRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_SendRequest.Unmarshal(m, b)
-}
-func (m *SendRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_SendRequest.Marshal(b, m, deterministic)
-}
-func (dst *SendRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_SendRequest.Merge(dst, src)
-}
-func (m *SendRequest) XXX_Size() int {
-	return xxx_messageInfo_SendRequest.Size(m)
-}
-func (m *SendRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_SendRequest.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{4}
 }
 
-var xxx_messageInfo_SendRequest proto.InternalMessageInfo
-
-func (m *SendRequest) GetMessage() string {
-	if m != nil {
-		return m.Message
+func (x *SendRequest) GetMessage() string {
+	if x != nil {
+		return x.Message
 	}
 	return ""
 }
 
-func (m *SendRequest) GetUserId() string {
-	if m != nil {
-		return m.UserId
+func (x *SendRequest) GetUserId() string {
+	if x != nil {
+		return x.UserId
 	}
 	return ""
 }
 
-func (m *SendRequest) GetChatId() string {
-	if m != nil {
-		return m.ChatId
+func (x *SendRequest) GetChatId() string {
+	if x != nil {
+		return x.ChatId
 	}
 	return ""
 }
 
-func (m *SendRequest) GetDateTime() *timestamp.Timestamp {
-	if m != nil {
-		return m.DateTime
+func (x *SendRequest) GetDateTime() *timestamp.Timestamp {
+	if x != nil {
+		return x.DateTime
 	}
 	return nil
 }
 
 type ReadReceipt struct {
-	UserId               string               `protobuf:"bytes,1,opt,name=userId,proto3" json:"userId,omitempty"`
-	ChatId               string               `protobuf:"bytes,2,opt,name=chatId,proto3" json:"chatId,omitempty"`
-	MessageId            string               `protobuf:"bytes,3,opt,name=messageId,proto3" json:"messageId,omitempty"`
-	DateTime             *timestamp.Timestamp `protobuf:"bytes,4,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
-	XXX_unrecognized     []byte               `json:"-"`
-	XXX_sizecache        int32                `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	UserId    string               `protobuf:"bytes,1,opt,name=userId,proto3" json:"userId,omitempty"`
+	ChatId    string               `protobuf:"bytes,2,opt,name=chatId,proto3" json:"chatId,omitempty"`
+	MessageId string               `protobuf:"bytes,3,opt,name=messageId,proto3" json:"messageId,omitempty"`
+	DateTime  *timestamp.Timestamp `protobuf:"bytes,4,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
 }
 
-func (m *ReadReceipt) Reset()         { *m = ReadReceipt{} }
-func (m *ReadReceipt) String() string { return proto.CompactTextString(m) }
-func (*ReadReceipt) ProtoMessage()    {}
+func (x *ReadReceipt) Reset() {
+	*x = ReadReceipt{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[5]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ReadReceipt) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ReadReceipt) ProtoMessage() {}
+
+func (x *ReadReceipt) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[5]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ReadReceipt.ProtoReflect.Descriptor instead.
 func (*ReadReceipt) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{5}
-}
-func (m *ReadReceipt) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_ReadReceipt.Unmarshal(m, b)
-}
-func (m *ReadReceipt) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_ReadReceipt.Marshal(b, m, deterministic)
-}
-func (dst *ReadReceipt) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_ReadReceipt.Merge(dst, src)
-}
-func (m *ReadReceipt) XXX_Size() int {
-	return xxx_messageInfo_ReadReceipt.Size(m)
-}
-func (m *ReadReceipt) XXX_DiscardUnknown() {
-	xxx_messageInfo_ReadReceipt.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{5}
 }
 
-var xxx_messageInfo_ReadReceipt proto.InternalMessageInfo
-
-func (m *ReadReceipt) GetUserId() string {
-	if m != nil {
-		return m.UserId
+func (x *ReadReceipt) GetUserId() string {
+	if x != nil {
+		return x.UserId
 	}
 	return ""
 }
 
-func (m *ReadReceipt) GetChatId() string {
-	if m != nil {
-		return m.ChatId
+func (x *ReadReceipt) GetChatId() string {
+	if x != nil {
+		return x.ChatId
 	}
 	return ""
 }
 
-func (m *ReadReceipt) GetMessageId() string {
-	if m != nil {
-		return m.MessageId
+func (x *ReadReceipt) GetMessageId() string {
+	if x != nil {
+		return x.MessageId
 	}
 	return ""
 }
 
-func (m *ReadReceipt) GetDateTime() *timestamp.Timestamp {
-	if m != nil {
-		return m.DateTime
+func (x *ReadReceipt) GetDateTime() *timestamp.Timestamp {
+	if x != nil {
+		return x.DateTime
 	}
 	return nil
 }
 
 type Message struct {
-	MessageId            string               `protobuf:"bytes,1,opt,name=messageId,proto3" json:"messageId,omitempty"`
-	SenderId             string               `protobuf:"bytes,2,opt,name=senderId,proto3" json:"senderId,omitempty"`
-	Type                 Message_MsgType      `protobuf:"varint,3,opt,name=type,proto3,enum=model.Message_MsgType" json:"type,omitempty"`
-	Text                 string               `protobuf:"bytes,4,opt,name=text,proto3" json:"text,omitempty"`
-	DateTime             *timestamp.Timestamp `protobuf:"bytes,5,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
-	XXX_unrecognized     []byte               `json:"-"`
-	XXX_sizecache        int32                `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	MessageId string               `protobuf:"bytes,1,opt,name=messageId,proto3" json:"messageId,omitempty"`
+	SenderId  string               `protobuf:"bytes,2,opt,name=senderId,proto3" json:"senderId,omitempty"`
+	Type      Message_MsgType      `protobuf:"varint,3,opt,name=type,proto3,enum=model.Message_MsgType" json:"type,omitempty"`
+	Text      string               `protobuf:"bytes,4,opt,name=text,proto3" json:"text,omitempty"`
+	DateTime  *timestamp.Timestamp `protobuf:"bytes,5,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
 }
 
-func (m *Message) Reset()         { *m = Message{} }
-func (m *Message) String() string { return proto.CompactTextString(m) }
-func (*Message) ProtoMessage()    {}
+func (x *Message) Reset() {
+	*x = Message{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_chat_proto_msgTypes[6]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Message) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Message) ProtoMessage() {}
+
+func (x *Message) ProtoReflect() protoreflect.Message {
+	mi := &file_chat_proto_msgTypes[6]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Message.ProtoReflect.Descriptor instead.
 func (*Message) Descriptor() ([]byte, []int) {
-	return fileDescriptor_chat_0c5614da34e0455c, []int{6}
-}
-func (m *Message) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_Message.Unmarshal(m, b)
-}
-func (m *Message) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_Message.Marshal(b, m, deterministic)
-}
-func (dst *Message) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Message.Merge(dst, src)
-}
-func (m *Message) XXX_Size() int {
-	return xxx_messageInfo_Message.Size(m)
-}
-func (m *Message) XXX_DiscardUnknown() {
-	xxx_messageInfo_Message.DiscardUnknown(m)
+	return file_chat_proto_rawDescGZIP(), []int{6}
 }
 
-var xxx_messageInfo_Message proto.InternalMessageInfo
-
-func (m *Message) GetMessageId() string {
-	if m != nil {
-		return m.MessageId
+func (x *Message) GetMessageId() string {
+	if x != nil {
+		return x.MessageId
 	}
 	return ""
 }
 
-func (m *Message) GetSenderId() string {
-	if m != nil {
-		return m.SenderId
+func (x *Message) GetSenderId() string {
+	if x != nil {
+		return x.SenderId
 	}
 	return ""
 }
 
-func (m *Message) GetType() Message_MsgType {
-	if m != nil {
-		return m.Type
+func (x *Message) GetType() Message_MsgType {
+	if x != nil {
+		return x.Type
 	}
 	return Message_READ
 }
 
-func (m *Message) GetText() string {
-	if m != nil {
-		return m.Text
+func (x *Message) GetText() string {
+	if x != nil {
+		return x.Text
 	}
 	return ""
 }
 
-func (m *Message) GetDateTime() *timestamp.Timestamp {
-	if m != nil {
-		return m.DateTime
+func (x *Message) GetDateTime() *timestamp.Timestamp {
+	if x != nil {
+		return x.DateTime
 	}
 	return nil
 }
 
-func init() {
-	proto.RegisterType((*AddRequest)(nil), "model.AddRequest")
-	proto.RegisterType((*LatestMessagesRequest)(nil), "model.LatestMessagesRequest")
-	proto.RegisterType((*PreviousMessagesRequest)(nil), "model.PreviousMessagesRequest")
-	proto.RegisterType((*MessageList)(nil), "model.MessageList")
-	proto.RegisterType((*SendRequest)(nil), "model.SendRequest")
-	proto.RegisterType((*ReadReceipt)(nil), "model.ReadReceipt")
-	proto.RegisterType((*Message)(nil), "model.Message")
-	proto.RegisterEnum("model.Message_MsgType", Message_MsgType_name, Message_MsgType_value)
+var File_chat_proto protoreflect.FileDescriptor
+
+var file_chat_proto_rawDesc = []byte{
+	0x0a, 0x0a, 0x63, 0x68, 0x61, 0x74, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x05, 0x6d, 0x6f,
+	0x64, 0x65, 0x6c, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74,
+	0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
+	0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74,
+	0x6f, 0x22, 0x3c, 0x0a, 0x0a, 0x41, 0x64, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
+	0x16, 0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x16, 0x0a, 0x06, 0x63, 0x68, 0x61, 0x74, 0x49,
+	0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x63, 0x68, 0x61, 0x74, 0x49, 0x64, 0x22,
+	0x2f, 0x0a, 0x15, 0x4c, 0x61, 0x74, 0x65, 0x73, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x16, 0x0a, 0x06, 0x63, 0x68, 0x61, 0x74,
+	0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x63, 0x68, 0x61, 0x74, 0x49, 0x64,
+	0x22, 0x5d, 0x0a, 0x17, 0x50, 0x72, 0x65, 0x76, 0x69, 0x6f, 0x75, 0x73, 0x4d, 0x65, 0x73, 0x73,
+	0x61, 0x67, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x16, 0x0a, 0x06, 0x63,
+	0x68, 0x61, 0x74, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x63, 0x68, 0x61,
+	0x74, 0x49, 0x64, 0x12, 0x2a, 0x0a, 0x10, 0x63, 0x75, 0x72, 0x72, 0x65, 0x6e, 0x74, 0x4d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x10, 0x63,
+	0x75, 0x72, 0x72, 0x65, 0x6e, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x64, 0x22,
+	0x39, 0x0a, 0x0b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x2a,
+	0x0a, 0x08, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b,
+	0x32, 0x0e, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x52, 0x08, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x22, 0x8f, 0x01, 0x0a, 0x0b, 0x53,
+	0x65, 0x6e, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x12, 0x16, 0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x02,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x16, 0x0a, 0x06,
+	0x63, 0x68, 0x61, 0x74, 0x49, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x63, 0x68,
+	0x61, 0x74, 0x49, 0x64, 0x12, 0x36, 0x0a, 0x08, 0x64, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65,
+	0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61,
+	0x6d, 0x70, 0x52, 0x08, 0x64, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x22, 0x93, 0x01, 0x0a,
+	0x0b, 0x52, 0x65, 0x61, 0x64, 0x52, 0x65, 0x63, 0x65, 0x69, 0x70, 0x74, 0x12, 0x16, 0x0a, 0x06,
+	0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x75, 0x73,
+	0x65, 0x72, 0x49, 0x64, 0x12, 0x16, 0x0a, 0x06, 0x63, 0x68, 0x61, 0x74, 0x49, 0x64, 0x18, 0x02,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x63, 0x68, 0x61, 0x74, 0x49, 0x64, 0x12, 0x1c, 0x0a, 0x09,
+	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x64, 0x12, 0x36, 0x0a, 0x08, 0x64, 0x61,
+	0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54,
+	0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x08, 0x64, 0x61, 0x74, 0x65, 0x54, 0x69,
+	0x6d, 0x65, 0x22, 0xdd, 0x01, 0x0a, 0x07, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1c,
+	0x0a, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x64, 0x12, 0x1a, 0x0a, 0x08,
+	0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08,
+	0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x49, 0x64, 0x12, 0x2a, 0x0a, 0x04, 0x74, 0x79, 0x70, 0x65,
+	0x18, 0x03, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x16, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x4d, 0x73, 0x67, 0x54, 0x79, 0x70, 0x65, 0x52, 0x04,
+	0x74, 0x79, 0x70, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x74, 0x65, 0x78, 0x74, 0x18, 0x04, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x04, 0x74, 0x65, 0x78, 0x74, 0x12, 0x36, 0x0a, 0x08, 0x64, 0x61, 0x74, 0x65,
+	0x54, 0x69, 0x6d, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f,
+	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d,
+	0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x08, 0x64, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65,
+	0x22, 0x20, 0x0a, 0x07, 0x4d, 0x73, 0x67, 0x54, 0x79, 0x70, 0x65, 0x12, 0x08, 0x0a, 0x04, 0x52,
+	0x45, 0x41, 0x44, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x4d, 0x45, 0x53, 0x53, 0x41, 0x47, 0x45,
+	0x10, 0x01, 0x32, 0x89, 0x04, 0x0a, 0x0b, 0x43, 0x68, 0x61, 0x74, 0x53, 0x65, 0x72, 0x76, 0x69,
+	0x63, 0x65, 0x12, 0x39, 0x0a, 0x0a, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x43, 0x68, 0x61, 0x74,
+	0x12, 0x11, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x41, 0x64, 0x64, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x00, 0x12, 0x37, 0x0a,
+	0x08, 0x4a, 0x6f, 0x69, 0x6e, 0x43, 0x68, 0x61, 0x74, 0x12, 0x11, 0x2e, 0x6d, 0x6f, 0x64, 0x65,
+	0x6c, 0x2e, 0x41, 0x64, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45,
+	0x6d, 0x70, 0x74, 0x79, 0x22, 0x00, 0x12, 0x38, 0x0a, 0x09, 0x4c, 0x65, 0x61, 0x76, 0x65, 0x43,
+	0x68, 0x61, 0x74, 0x12, 0x11, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x41, 0x64, 0x64, 0x52,
+	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x00,
+	0x12, 0x47, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x4c, 0x61, 0x74, 0x65, 0x73, 0x74, 0x4d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x73, 0x12, 0x1c, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4c, 0x61,
+	0x74, 0x65, 0x73, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x12, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4d, 0x65, 0x73, 0x73,
+	0x61, 0x67, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x22, 0x00, 0x12, 0x4b, 0x0a, 0x13, 0x47, 0x65, 0x74,
+	0x50, 0x72, 0x65, 0x76, 0x69, 0x6f, 0x75, 0x73, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73,
+	0x12, 0x1e, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x50, 0x72, 0x65, 0x76, 0x69, 0x6f, 0x75,
+	0x73, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x1a, 0x12, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x4c, 0x69, 0x73, 0x74, 0x22, 0x00, 0x12, 0x49, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x52, 0x65, 0x63,
+	0x65, 0x6e, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x12, 0x1e, 0x2e, 0x6d, 0x6f,
+	0x64, 0x65, 0x6c, 0x2e, 0x50, 0x72, 0x65, 0x76, 0x69, 0x6f, 0x75, 0x73, 0x4d, 0x65, 0x73, 0x73,
+	0x61, 0x67, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x12, 0x2e, 0x6d, 0x6f,
+	0x64, 0x65, 0x6c, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x22,
+	0x00, 0x12, 0x34, 0x0a, 0x04, 0x53, 0x65, 0x6e, 0x64, 0x12, 0x12, 0x2e, 0x6d, 0x6f, 0x64, 0x65,
+	0x6c, 0x2e, 0x53, 0x65, 0x6e, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e,
+	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
+	0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x00, 0x12, 0x35, 0x0a, 0x09, 0x53, 0x75, 0x62, 0x73, 0x63,
+	0x72, 0x69, 0x62, 0x65, 0x12, 0x12, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x52, 0x65, 0x61,
+	0x64, 0x52, 0x65, 0x63, 0x65, 0x69, 0x70, 0x74, 0x1a, 0x0e, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c,
+	0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x00, 0x28, 0x01, 0x30, 0x01, 0x42, 0x24,
+	0x0a, 0x20, 0x63, 0x6f, 0x6d, 0x2e, 0x63, 0x73, 0x68, 0x65, 0x70, 0x34, 0x2e, 0x70, 0x72, 0x65,
+	0x6d, 0x69, 0x65, 0x72, 0x70, 0x72, 0x65, 0x64, 0x69, 0x63, 0x74, 0x6f, 0x72, 0x2e, 0x63, 0x68,
+	0x61, 0x74, 0x50, 0x01, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+}
+
+var (
+	file_chat_proto_rawDescOnce sync.Once
+	file_chat_proto_rawDescData = file_chat_proto_rawDesc
+)
+
+func file_chat_proto_rawDescGZIP() []byte {
+	file_chat_proto_rawDescOnce.Do(func() {
+		file_chat_proto_rawDescData = protoimpl.X.CompressGZIP(file_chat_proto_rawDescData)
+	})
+	return file_chat_proto_rawDescData
+}
+
+var file_chat_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_chat_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
+var file_chat_proto_goTypes = []interface{}{
+	(Message_MsgType)(0),            // 0: model.Message.MsgType
+	(*AddRequest)(nil),              // 1: model.AddRequest
+	(*LatestMessagesRequest)(nil),   // 2: model.LatestMessagesRequest
+	(*PreviousMessagesRequest)(nil), // 3: model.PreviousMessagesRequest
+	(*MessageList)(nil),             // 4: model.MessageList
+	(*SendRequest)(nil),             // 5: model.SendRequest
+	(*ReadReceipt)(nil),             // 6: model.ReadReceipt
+	(*Message)(nil),                 // 7: model.Message
+	(*timestamp.Timestamp)(nil),     // 8: google.protobuf.Timestamp
+	(*empty.Empty)(nil),             // 9: google.protobuf.Empty
+}
+var file_chat_proto_depIdxs = []int32{
+	7,  // 0: model.MessageList.messages:type_name -> model.Message
+	8,  // 1: model.SendRequest.dateTime:type_name -> google.protobuf.Timestamp
+	8,  // 2: model.ReadReceipt.dateTime:type_name -> google.protobuf.Timestamp
+	0,  // 3: model.Message.type:type_name -> model.Message.MsgType
+	8,  // 4: model.Message.dateTime:type_name -> google.protobuf.Timestamp
+	1,  // 5: model.ChatService.CreateChat:input_type -> model.AddRequest
+	1,  // 6: model.ChatService.JoinChat:input_type -> model.AddRequest
+	1,  // 7: model.ChatService.LeaveChat:input_type -> model.AddRequest
+	2,  // 8: model.ChatService.GetLatestMessages:input_type -> model.LatestMessagesRequest
+	3,  // 9: model.ChatService.GetPreviousMessages:input_type -> model.PreviousMessagesRequest
+	3,  // 10: model.ChatService.GetRecentMessages:input_type -> model.PreviousMessagesRequest
+	5,  // 11: model.ChatService.Send:input_type -> model.SendRequest
+	6,  // 12: model.ChatService.Subscribe:input_type -> model.ReadReceipt
+	9,  // 13: model.ChatService.CreateChat:output_type -> google.protobuf.Empty
+	9,  // 14: model.ChatService.JoinChat:output_type -> google.protobuf.Empty
+	9,  // 15: model.ChatService.LeaveChat:output_type -> google.protobuf.Empty
+	4,  // 16: model.ChatService.GetLatestMessages:output_type -> model.MessageList
+	4,  // 17: model.ChatService.GetPreviousMessages:output_type -> model.MessageList
+	4,  // 18: model.ChatService.GetRecentMessages:output_type -> model.MessageList
+	9,  // 19: model.ChatService.Send:output_type -> google.protobuf.Empty
+	7,  // 20: model.ChatService.Subscribe:output_type -> model.Message
+	13, // [13:21] is the sub-list for method output_type
+	5,  // [5:13] is the sub-list for method input_type
+	5,  // [5:5] is the sub-list for extension type_name
+	5,  // [5:5] is the sub-list for extension extendee
+	0,  // [0:5] is the sub-list for field type_name
+}
+
+func init() { file_chat_proto_init() }
+func file_chat_proto_init() {
+	if File_chat_proto != nil {
+		return
+	}
+	if !protoimpl.UnsafeEnabled {
+		file_chat_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*AddRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_chat_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*LatestMessagesRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_chat_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*PreviousMessagesRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_chat_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*MessageList); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_chat_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*SendRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_chat_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*ReadReceipt); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_chat_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Message); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+	}
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: file_chat_proto_rawDesc,
+			NumEnums:      1,
+			NumMessages:   7,
+			NumExtensions: 0,
+			NumServices:   1,
+		},
+		GoTypes:           file_chat_proto_goTypes,
+		DependencyIndexes: file_chat_proto_depIdxs,
+		EnumInfos:         file_chat_proto_enumTypes,
+		MessageInfos:      file_chat_proto_msgTypes,
+	}.Build()
+	File_chat_proto = out.File
+	file_chat_proto_rawDesc = nil
+	file_chat_proto_goTypes = nil
+	file_chat_proto_depIdxs = nil
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ context.Context
-var _ grpc.ClientConn
+var _ grpc.ClientConnInterface
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion4
+const _ = grpc.SupportPackageIsVersion6
 
 // ChatServiceClient is the client API for ChatService service.
 //
@@ -444,10 +787,10 @@ type ChatServiceClient interface {
 }
 
 type chatServiceClient struct {
-	cc *grpc.ClientConn
+	cc grpc.ClientConnInterface
 }
 
-func NewChatServiceClient(cc *grpc.ClientConn) ChatServiceClient {
+func NewChatServiceClient(cc grpc.ClientConnInterface) ChatServiceClient {
 	return &chatServiceClient{cc}
 }
 
@@ -555,6 +898,35 @@ type ChatServiceServer interface {
 	GetRecentMessages(context.Context, *PreviousMessagesRequest) (*MessageList, error)
 	Send(context.Context, *SendRequest) (*empty.Empty, error)
 	Subscribe(ChatService_SubscribeServer) error
+}
+
+// UnimplementedChatServiceServer can be embedded to have forward compatible implementations.
+type UnimplementedChatServiceServer struct {
+}
+
+func (*UnimplementedChatServiceServer) CreateChat(context.Context, *AddRequest) (*empty.Empty, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method CreateChat not implemented")
+}
+func (*UnimplementedChatServiceServer) JoinChat(context.Context, *AddRequest) (*empty.Empty, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method JoinChat not implemented")
+}
+func (*UnimplementedChatServiceServer) LeaveChat(context.Context, *AddRequest) (*empty.Empty, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method LeaveChat not implemented")
+}
+func (*UnimplementedChatServiceServer) GetLatestMessages(context.Context, *LatestMessagesRequest) (*MessageList, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetLatestMessages not implemented")
+}
+func (*UnimplementedChatServiceServer) GetPreviousMessages(context.Context, *PreviousMessagesRequest) (*MessageList, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetPreviousMessages not implemented")
+}
+func (*UnimplementedChatServiceServer) GetRecentMessages(context.Context, *PreviousMessagesRequest) (*MessageList, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetRecentMessages not implemented")
+}
+func (*UnimplementedChatServiceServer) Send(context.Context, *SendRequest) (*empty.Empty, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method Send not implemented")
+}
+func (*UnimplementedChatServiceServer) Subscribe(ChatService_SubscribeServer) error {
+	return status.Errorf(codes.Unimplemented, "method Subscribe not implemented")
 }
 
 func RegisterChatServiceServer(s *grpc.Server, srv ChatServiceServer) {
@@ -755,45 +1127,4 @@ var _ChatService_serviceDesc = grpc.ServiceDesc{
 		},
 	},
 	Metadata: "chat.proto",
-}
-
-func init() { proto.RegisterFile("chat.proto", fileDescriptor_chat_0c5614da34e0455c) }
-
-var fileDescriptor_chat_0c5614da34e0455c = []byte{
-	// 555 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x54, 0xcd, 0x6e, 0xda, 0x4c,
-	0x14, 0x65, 0x12, 0x27, 0xc0, 0xb5, 0x14, 0x91, 0xf9, 0xf4, 0x51, 0xe4, 0x46, 0x2d, 0xb2, 0xba,
-	0x40, 0x2c, 0x86, 0x8a, 0xa6, 0x3f, 0x91, 0xba, 0xa1, 0x29, 0x42, 0xb4, 0x20, 0x45, 0x86, 0x6d,
-	0x17, 0xc6, 0xbe, 0x85, 0x91, 0x62, 0xec, 0xce, 0x8c, 0x51, 0x79, 0x84, 0xae, 0xba, 0xe8, 0xf3,
-	0xf5, 0x5d, 0x2a, 0xdb, 0x83, 0x83, 0x49, 0xa9, 0x4a, 0xba, 0xf3, 0xfd, 0x3d, 0xc7, 0xe7, 0xfa,
-	0x18, 0xc0, 0x5b, 0xb8, 0x8a, 0x45, 0x22, 0x54, 0x21, 0x3d, 0x09, 0x42, 0x1f, 0x6f, 0xad, 0xc7,
-	0xf3, 0x30, 0x9c, 0xdf, 0x62, 0x27, 0x4d, 0xce, 0xe2, 0xcf, 0x1d, 0x0c, 0x22, 0xb5, 0xce, 0x7a,
-	0xac, 0xa7, 0xbb, 0x45, 0xc5, 0x03, 0x94, 0xca, 0x0d, 0xa2, 0xac, 0xc1, 0x7e, 0x0b, 0xd0, 0xf3,
-	0x7d, 0x07, 0xbf, 0xc4, 0x28, 0x15, 0xad, 0xc3, 0x69, 0x2c, 0x51, 0x0c, 0xfd, 0x06, 0x69, 0x92,
-	0x56, 0xd5, 0xd1, 0x51, 0x92, 0x4f, 0x80, 0x87, 0x7e, 0xe3, 0x28, 0xcb, 0x67, 0x91, 0xdd, 0x81,
-	0xff, 0x47, 0xae, 0x42, 0xa9, 0xc6, 0x28, 0xa5, 0x3b, 0x47, 0xb9, 0xb5, 0x48, 0x0f, 0x90, 0xc2,
-	0xc0, 0x27, 0x78, 0x74, 0x23, 0x70, 0xc5, 0xc3, 0x58, 0xfe, 0xe5, 0x08, 0x6d, 0x43, 0xcd, 0x8b,
-	0x85, 0xc0, 0xe5, 0x06, 0x24, 0x67, 0x71, 0x2f, 0x6f, 0x5f, 0x81, 0xa9, 0x83, 0x11, 0x97, 0x8a,
-	0xb6, 0xa1, 0x12, 0x68, 0x94, 0x06, 0x69, 0x1e, 0xb7, 0xcc, 0xee, 0x19, 0x4b, 0x45, 0x63, 0xba,
-	0xcb, 0xc9, 0xeb, 0xf6, 0x77, 0x02, 0xe6, 0x04, 0x97, 0xb9, 0x14, 0x0d, 0x28, 0xeb, 0x9a, 0xe6,
-	0xb3, 0x09, 0xb7, 0x44, 0x3a, 0xda, 0x23, 0xd2, 0x71, 0xe1, 0x05, 0x5e, 0x41, 0xc5, 0x77, 0x15,
-	0x4e, 0x79, 0x80, 0x0d, 0xa3, 0x49, 0x5a, 0x66, 0xd7, 0x62, 0xd9, 0x59, 0xd8, 0xe6, 0x2c, 0x6c,
-	0xba, 0x39, 0x8b, 0x93, 0xf7, 0xda, 0x3f, 0x08, 0x98, 0x0e, 0xba, 0xbe, 0x83, 0x1e, 0xf2, 0xe8,
-	0xe0, 0xe3, 0xd0, 0x0b, 0xa8, 0x06, 0xb9, 0x62, 0x19, 0xa5, 0xbb, 0xc4, 0x83, 0x59, 0xfd, 0x24,
-	0x50, 0xd6, 0xea, 0x15, 0x11, 0xc8, 0x2e, 0x82, 0x05, 0x15, 0x89, 0x4b, 0x7f, 0x4b, 0xa9, 0x3c,
-	0xa6, 0x6d, 0x30, 0xd4, 0x3a, 0xc2, 0x94, 0xd6, 0x59, 0xb7, 0x5e, 0xbc, 0x0a, 0x1b, 0xcb, 0xf9,
-	0x74, 0x1d, 0xa1, 0x93, 0xf6, 0x50, 0x0a, 0x86, 0xc2, 0xaf, 0x2a, 0x65, 0x59, 0x75, 0xd2, 0xe7,
-	0x02, 0xfb, 0x93, 0x03, 0xd8, 0x37, 0xa1, 0xac, 0x97, 0xd3, 0x0a, 0x18, 0x4e, 0xbf, 0xf7, 0xbe,
-	0x56, 0xa2, 0x26, 0x94, 0xc7, 0xfd, 0xc9, 0xa4, 0x37, 0xe8, 0xd7, 0x48, 0xf7, 0x9b, 0x01, 0xe6,
-	0xf5, 0xc2, 0x55, 0x13, 0x14, 0x2b, 0xee, 0x21, 0xbd, 0x02, 0xb8, 0x16, 0xe8, 0x2a, 0x4c, 0x92,
-	0xf4, 0x5c, 0x33, 0xbd, 0xf3, 0x8c, 0x55, 0xbf, 0x07, 0xdc, 0x4f, 0x0c, 0x68, 0x97, 0xe8, 0x6b,
-	0xa8, 0x7c, 0x08, 0xf9, 0xf2, 0xf0, 0xc1, 0x37, 0x50, 0x1d, 0xa1, 0xbb, 0x7a, 0x00, 0xe4, 0x00,
-	0xce, 0x07, 0xa8, 0x8a, 0x9e, 0xa4, 0x17, 0x7a, 0xc3, 0x6f, 0xad, 0x6a, 0xd1, 0xa2, 0xf8, 0x89,
-	0x71, 0xec, 0x12, 0xfd, 0x08, 0xff, 0x0d, 0x50, 0xed, 0x7a, 0x95, 0x3e, 0xd1, 0xcd, 0x7b, 0x4c,
-	0xbc, 0x67, 0xd9, 0x30, 0x65, 0x95, 0x7c, 0xc7, 0x4b, 0xf5, 0x8f, 0xab, 0x2e, 0xc1, 0x48, 0x5c,
-	0x4a, 0x37, 0xd5, 0x2d, 0xcb, 0xfe, 0x41, 0x96, 0x97, 0x50, 0x9d, 0xc4, 0x33, 0xe9, 0x09, 0x3e,
-	0xc3, 0x7c, 0x74, 0xcb, 0x5b, 0xd6, 0xce, 0x7f, 0xc1, 0x2e, 0xb5, 0xc8, 0x73, 0xf2, 0xee, 0x19,
-	0x34, 0xbd, 0x30, 0x60, 0x9e, 0x5c, 0x60, 0x74, 0xc9, 0x22, 0x81, 0x01, 0x47, 0x11, 0x09, 0xf4,
-	0xb9, 0xa7, 0x42, 0xc1, 0x12, 0x9f, 0xdd, 0x90, 0xd9, 0x69, 0x0a, 0xf7, 0xe2, 0x57, 0x00, 0x00,
-	0x00, 0xff, 0xff, 0x69, 0xd6, 0xfb, 0x66, 0x9c, 0x05, 0x00, 0x00,
 }
