@@ -1,30 +1,35 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.25.0
+// 	protoc        v3.12.3
 // source: fixture.proto
 
 package model
 
 import (
 	context "context"
-	fmt "fmt"
 	proto "github.com/golang/protobuf/proto"
 	empty "github.com/golang/protobuf/ptypes/empty"
 	timestamp "github.com/golang/protobuf/ptypes/timestamp"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
-	math "math"
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
 )
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+// This is a compile-time assertion that a sufficiently up-to-date version
+// of the legacy proto package is being used.
+const _ = proto.ProtoPackageIsVersion4
 
 type TeamMatchResult_Result int32
 
@@ -34,24 +39,45 @@ const (
 	TeamMatchResult_LOSS TeamMatchResult_Result = 2
 )
 
-var TeamMatchResult_Result_name = map[int32]string{
-	0: "WIN",
-	1: "DRAW",
-	2: "LOSS",
-}
+// Enum value maps for TeamMatchResult_Result.
+var (
+	TeamMatchResult_Result_name = map[int32]string{
+		0: "WIN",
+		1: "DRAW",
+		2: "LOSS",
+	}
+	TeamMatchResult_Result_value = map[string]int32{
+		"WIN":  0,
+		"DRAW": 1,
+		"LOSS": 2,
+	}
+)
 
-var TeamMatchResult_Result_value = map[string]int32{
-	"WIN":  0,
-	"DRAW": 1,
-	"LOSS": 2,
+func (x TeamMatchResult_Result) Enum() *TeamMatchResult_Result {
+	p := new(TeamMatchResult_Result)
+	*p = x
+	return p
 }
 
 func (x TeamMatchResult_Result) String() string {
-	return proto.EnumName(TeamMatchResult_Result_name, int32(x))
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
 }
 
+func (TeamMatchResult_Result) Descriptor() protoreflect.EnumDescriptor {
+	return file_fixture_proto_enumTypes[0].Descriptor()
+}
+
+func (TeamMatchResult_Result) Type() protoreflect.EnumType {
+	return &file_fixture_proto_enumTypes[0]
+}
+
+func (x TeamMatchResult_Result) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use TeamMatchResult_Result.Descriptor instead.
 func (TeamMatchResult_Result) EnumDescriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{4, 0}
+	return file_fixture_proto_rawDescGZIP(), []int{4, 0}
 }
 
 type TeamMatchResult_Location int32
@@ -61,402 +87,644 @@ const (
 	TeamMatchResult_AWAY TeamMatchResult_Location = 1
 )
 
-var TeamMatchResult_Location_name = map[int32]string{
-	0: "HOME",
-	1: "AWAY",
-}
+// Enum value maps for TeamMatchResult_Location.
+var (
+	TeamMatchResult_Location_name = map[int32]string{
+		0: "HOME",
+		1: "AWAY",
+	}
+	TeamMatchResult_Location_value = map[string]int32{
+		"HOME": 0,
+		"AWAY": 1,
+	}
+)
 
-var TeamMatchResult_Location_value = map[string]int32{
-	"HOME": 0,
-	"AWAY": 1,
+func (x TeamMatchResult_Location) Enum() *TeamMatchResult_Location {
+	p := new(TeamMatchResult_Location)
+	*p = x
+	return p
 }
 
 func (x TeamMatchResult_Location) String() string {
-	return proto.EnumName(TeamMatchResult_Location_name, int32(x))
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
 }
 
+func (TeamMatchResult_Location) Descriptor() protoreflect.EnumDescriptor {
+	return file_fixture_proto_enumTypes[1].Descriptor()
+}
+
+func (TeamMatchResult_Location) Type() protoreflect.EnumType {
+	return &file_fixture_proto_enumTypes[1]
+}
+
+func (x TeamMatchResult_Location) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use TeamMatchResult_Location.Descriptor instead.
 func (TeamMatchResult_Location) EnumDescriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{4, 1}
+	return file_fixture_proto_rawDescGZIP(), []int{4, 1}
 }
 
 type Matches struct {
-	Matches              []*Match `protobuf:"bytes,1,rep,name=matches,proto3" json:"matches,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Matches []*Match `protobuf:"bytes,1,rep,name=matches,proto3" json:"matches,omitempty"`
 }
 
-func (m *Matches) Reset()         { *m = Matches{} }
-func (m *Matches) String() string { return proto.CompactTextString(m) }
-func (*Matches) ProtoMessage()    {}
+func (x *Matches) Reset() {
+	*x = Matches{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fixture_proto_msgTypes[0]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Matches) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Matches) ProtoMessage() {}
+
+func (x *Matches) ProtoReflect() protoreflect.Message {
+	mi := &file_fixture_proto_msgTypes[0]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Matches.ProtoReflect.Descriptor instead.
 func (*Matches) Descriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{0}
+	return file_fixture_proto_rawDescGZIP(), []int{0}
 }
 
-func (m *Matches) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_Matches.Unmarshal(m, b)
-}
-func (m *Matches) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_Matches.Marshal(b, m, deterministic)
-}
-func (m *Matches) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Matches.Merge(m, src)
-}
-func (m *Matches) XXX_Size() int {
-	return xxx_messageInfo_Matches.Size(m)
-}
-func (m *Matches) XXX_DiscardUnknown() {
-	xxx_messageInfo_Matches.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_Matches proto.InternalMessageInfo
-
-func (m *Matches) GetMatches() []*Match {
-	if m != nil {
-		return m.Matches
+func (x *Matches) GetMatches() []*Match {
+	if x != nil {
+		return x.Matches
 	}
 	return nil
 }
 
 type Match struct {
-	Id                   string               `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	HTeam                string               `protobuf:"bytes,2,opt,name=hTeam,proto3" json:"hTeam,omitempty"`
-	ATeam                string               `protobuf:"bytes,3,opt,name=aTeam,proto3" json:"aTeam,omitempty"`
-	HGoals               int32                `protobuf:"varint,4,opt,name=hGoals,proto3" json:"hGoals,omitempty"`
-	AGoals               int32                `protobuf:"varint,5,opt,name=aGoals,proto3" json:"aGoals,omitempty"`
-	Played               int32                `protobuf:"varint,6,opt,name=played,proto3" json:"played,omitempty"`
-	DateTime             *timestamp.Timestamp `protobuf:"bytes,7,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
-	Matchday             int32                `protobuf:"varint,8,opt,name=matchday,proto3" json:"matchday,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
-	XXX_unrecognized     []byte               `json:"-"`
-	XXX_sizecache        int32                `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Id       string               `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	HTeam    string               `protobuf:"bytes,2,opt,name=hTeam,proto3" json:"hTeam,omitempty"`
+	ATeam    string               `protobuf:"bytes,3,opt,name=aTeam,proto3" json:"aTeam,omitempty"`
+	HGoals   int32                `protobuf:"varint,4,opt,name=hGoals,proto3" json:"hGoals,omitempty"`
+	AGoals   int32                `protobuf:"varint,5,opt,name=aGoals,proto3" json:"aGoals,omitempty"`
+	Played   int32                `protobuf:"varint,6,opt,name=played,proto3" json:"played,omitempty"`
+	DateTime *timestamp.Timestamp `protobuf:"bytes,7,opt,name=dateTime,proto3" json:"dateTime,omitempty"`
+	Matchday int32                `protobuf:"varint,8,opt,name=matchday,proto3" json:"matchday,omitempty"`
 }
 
-func (m *Match) Reset()         { *m = Match{} }
-func (m *Match) String() string { return proto.CompactTextString(m) }
-func (*Match) ProtoMessage()    {}
+func (x *Match) Reset() {
+	*x = Match{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fixture_proto_msgTypes[1]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Match) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Match) ProtoMessage() {}
+
+func (x *Match) ProtoReflect() protoreflect.Message {
+	mi := &file_fixture_proto_msgTypes[1]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Match.ProtoReflect.Descriptor instead.
 func (*Match) Descriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{1}
+	return file_fixture_proto_rawDescGZIP(), []int{1}
 }
 
-func (m *Match) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_Match.Unmarshal(m, b)
-}
-func (m *Match) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_Match.Marshal(b, m, deterministic)
-}
-func (m *Match) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Match.Merge(m, src)
-}
-func (m *Match) XXX_Size() int {
-	return xxx_messageInfo_Match.Size(m)
-}
-func (m *Match) XXX_DiscardUnknown() {
-	xxx_messageInfo_Match.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_Match proto.InternalMessageInfo
-
-func (m *Match) GetId() string {
-	if m != nil {
-		return m.Id
+func (x *Match) GetId() string {
+	if x != nil {
+		return x.Id
 	}
 	return ""
 }
 
-func (m *Match) GetHTeam() string {
-	if m != nil {
-		return m.HTeam
+func (x *Match) GetHTeam() string {
+	if x != nil {
+		return x.HTeam
 	}
 	return ""
 }
 
-func (m *Match) GetATeam() string {
-	if m != nil {
-		return m.ATeam
+func (x *Match) GetATeam() string {
+	if x != nil {
+		return x.ATeam
 	}
 	return ""
 }
 
-func (m *Match) GetHGoals() int32 {
-	if m != nil {
-		return m.HGoals
+func (x *Match) GetHGoals() int32 {
+	if x != nil {
+		return x.HGoals
 	}
 	return 0
 }
 
-func (m *Match) GetAGoals() int32 {
-	if m != nil {
-		return m.AGoals
+func (x *Match) GetAGoals() int32 {
+	if x != nil {
+		return x.AGoals
 	}
 	return 0
 }
 
-func (m *Match) GetPlayed() int32 {
-	if m != nil {
-		return m.Played
+func (x *Match) GetPlayed() int32 {
+	if x != nil {
+		return x.Played
 	}
 	return 0
 }
 
-func (m *Match) GetDateTime() *timestamp.Timestamp {
-	if m != nil {
-		return m.DateTime
+func (x *Match) GetDateTime() *timestamp.Timestamp {
+	if x != nil {
+		return x.DateTime
 	}
 	return nil
 }
 
-func (m *Match) GetMatchday() int32 {
-	if m != nil {
-		return m.Matchday
+func (x *Match) GetMatchday() int32 {
+	if x != nil {
+		return x.Matchday
 	}
 	return 0
 }
 
 type Forms struct {
-	Teams                map[string]*TeamForm `protobuf:"bytes,1,rep,name=teams,proto3" json:"teams,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
-	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
-	XXX_unrecognized     []byte               `json:"-"`
-	XXX_sizecache        int32                `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Teams map[string]*TeamForm `protobuf:"bytes,1,rep,name=teams,proto3" json:"teams,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
 }
 
-func (m *Forms) Reset()         { *m = Forms{} }
-func (m *Forms) String() string { return proto.CompactTextString(m) }
-func (*Forms) ProtoMessage()    {}
+func (x *Forms) Reset() {
+	*x = Forms{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fixture_proto_msgTypes[2]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Forms) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Forms) ProtoMessage() {}
+
+func (x *Forms) ProtoReflect() protoreflect.Message {
+	mi := &file_fixture_proto_msgTypes[2]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Forms.ProtoReflect.Descriptor instead.
 func (*Forms) Descriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{2}
+	return file_fixture_proto_rawDescGZIP(), []int{2}
 }
 
-func (m *Forms) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_Forms.Unmarshal(m, b)
-}
-func (m *Forms) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_Forms.Marshal(b, m, deterministic)
-}
-func (m *Forms) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Forms.Merge(m, src)
-}
-func (m *Forms) XXX_Size() int {
-	return xxx_messageInfo_Forms.Size(m)
-}
-func (m *Forms) XXX_DiscardUnknown() {
-	xxx_messageInfo_Forms.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_Forms proto.InternalMessageInfo
-
-func (m *Forms) GetTeams() map[string]*TeamForm {
-	if m != nil {
-		return m.Teams
+func (x *Forms) GetTeams() map[string]*TeamForm {
+	if x != nil {
+		return x.Teams
 	}
 	return nil
 }
 
 type TeamForm struct {
-	Forms                []*TeamMatchResult `protobuf:"bytes,1,rep,name=forms,proto3" json:"forms,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}           `json:"-"`
-	XXX_unrecognized     []byte             `json:"-"`
-	XXX_sizecache        int32              `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Forms []*TeamMatchResult `protobuf:"bytes,1,rep,name=forms,proto3" json:"forms,omitempty"`
 }
 
-func (m *TeamForm) Reset()         { *m = TeamForm{} }
-func (m *TeamForm) String() string { return proto.CompactTextString(m) }
-func (*TeamForm) ProtoMessage()    {}
+func (x *TeamForm) Reset() {
+	*x = TeamForm{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fixture_proto_msgTypes[3]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TeamForm) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TeamForm) ProtoMessage() {}
+
+func (x *TeamForm) ProtoReflect() protoreflect.Message {
+	mi := &file_fixture_proto_msgTypes[3]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TeamForm.ProtoReflect.Descriptor instead.
 func (*TeamForm) Descriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{3}
+	return file_fixture_proto_rawDescGZIP(), []int{3}
 }
 
-func (m *TeamForm) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_TeamForm.Unmarshal(m, b)
-}
-func (m *TeamForm) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_TeamForm.Marshal(b, m, deterministic)
-}
-func (m *TeamForm) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TeamForm.Merge(m, src)
-}
-func (m *TeamForm) XXX_Size() int {
-	return xxx_messageInfo_TeamForm.Size(m)
-}
-func (m *TeamForm) XXX_DiscardUnknown() {
-	xxx_messageInfo_TeamForm.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TeamForm proto.InternalMessageInfo
-
-func (m *TeamForm) GetForms() []*TeamMatchResult {
-	if m != nil {
-		return m.Forms
+func (x *TeamForm) GetForms() []*TeamMatchResult {
+	if x != nil {
+		return x.Forms
 	}
 	return nil
 }
 
 type TeamMatchResult struct {
-	Result               TeamMatchResult_Result   `protobuf:"varint,1,opt,name=result,proto3,enum=model.TeamMatchResult_Result" json:"result,omitempty"`
-	Score                string                   `protobuf:"bytes,2,opt,name=score,proto3" json:"score,omitempty"`
-	Opponent             string                   `protobuf:"bytes,3,opt,name=opponent,proto3" json:"opponent,omitempty"`
-	Location             TeamMatchResult_Location `protobuf:"varint,4,opt,name=location,proto3,enum=model.TeamMatchResult_Location" json:"location,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}                 `json:"-"`
-	XXX_unrecognized     []byte                   `json:"-"`
-	XXX_sizecache        int32                    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Result   TeamMatchResult_Result   `protobuf:"varint,1,opt,name=result,proto3,enum=model.TeamMatchResult_Result" json:"result,omitempty"`
+	Score    string                   `protobuf:"bytes,2,opt,name=score,proto3" json:"score,omitempty"`
+	Opponent string                   `protobuf:"bytes,3,opt,name=opponent,proto3" json:"opponent,omitempty"`
+	Location TeamMatchResult_Location `protobuf:"varint,4,opt,name=location,proto3,enum=model.TeamMatchResult_Location" json:"location,omitempty"`
 }
 
-func (m *TeamMatchResult) Reset()         { *m = TeamMatchResult{} }
-func (m *TeamMatchResult) String() string { return proto.CompactTextString(m) }
-func (*TeamMatchResult) ProtoMessage()    {}
+func (x *TeamMatchResult) Reset() {
+	*x = TeamMatchResult{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fixture_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TeamMatchResult) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TeamMatchResult) ProtoMessage() {}
+
+func (x *TeamMatchResult) ProtoReflect() protoreflect.Message {
+	mi := &file_fixture_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TeamMatchResult.ProtoReflect.Descriptor instead.
 func (*TeamMatchResult) Descriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{4}
+	return file_fixture_proto_rawDescGZIP(), []int{4}
 }
 
-func (m *TeamMatchResult) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_TeamMatchResult.Unmarshal(m, b)
-}
-func (m *TeamMatchResult) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_TeamMatchResult.Marshal(b, m, deterministic)
-}
-func (m *TeamMatchResult) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TeamMatchResult.Merge(m, src)
-}
-func (m *TeamMatchResult) XXX_Size() int {
-	return xxx_messageInfo_TeamMatchResult.Size(m)
-}
-func (m *TeamMatchResult) XXX_DiscardUnknown() {
-	xxx_messageInfo_TeamMatchResult.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TeamMatchResult proto.InternalMessageInfo
-
-func (m *TeamMatchResult) GetResult() TeamMatchResult_Result {
-	if m != nil {
-		return m.Result
+func (x *TeamMatchResult) GetResult() TeamMatchResult_Result {
+	if x != nil {
+		return x.Result
 	}
 	return TeamMatchResult_WIN
 }
 
-func (m *TeamMatchResult) GetScore() string {
-	if m != nil {
-		return m.Score
+func (x *TeamMatchResult) GetScore() string {
+	if x != nil {
+		return x.Score
 	}
 	return ""
 }
 
-func (m *TeamMatchResult) GetOpponent() string {
-	if m != nil {
-		return m.Opponent
+func (x *TeamMatchResult) GetOpponent() string {
+	if x != nil {
+		return x.Opponent
 	}
 	return ""
 }
 
-func (m *TeamMatchResult) GetLocation() TeamMatchResult_Location {
-	if m != nil {
-		return m.Location
+func (x *TeamMatchResult) GetLocation() TeamMatchResult_Location {
+	if x != nil {
+		return x.Location
 	}
 	return TeamMatchResult_HOME
 }
 
 type FutureMatches struct {
-	Matches              map[string]string `protobuf:"bytes,1,rep,name=matches,proto3" json:"matches,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
-	XXX_NoUnkeyedLiteral struct{}          `json:"-"`
-	XXX_unrecognized     []byte            `json:"-"`
-	XXX_sizecache        int32             `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Matches map[string]string `protobuf:"bytes,1,rep,name=matches,proto3" json:"matches,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
 }
 
-func (m *FutureMatches) Reset()         { *m = FutureMatches{} }
-func (m *FutureMatches) String() string { return proto.CompactTextString(m) }
-func (*FutureMatches) ProtoMessage()    {}
+func (x *FutureMatches) Reset() {
+	*x = FutureMatches{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fixture_proto_msgTypes[5]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *FutureMatches) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*FutureMatches) ProtoMessage() {}
+
+func (x *FutureMatches) ProtoReflect() protoreflect.Message {
+	mi := &file_fixture_proto_msgTypes[5]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use FutureMatches.ProtoReflect.Descriptor instead.
 func (*FutureMatches) Descriptor() ([]byte, []int) {
-	return fileDescriptor_d81820e403a865f4, []int{5}
+	return file_fixture_proto_rawDescGZIP(), []int{5}
 }
 
-func (m *FutureMatches) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_FutureMatches.Unmarshal(m, b)
-}
-func (m *FutureMatches) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_FutureMatches.Marshal(b, m, deterministic)
-}
-func (m *FutureMatches) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_FutureMatches.Merge(m, src)
-}
-func (m *FutureMatches) XXX_Size() int {
-	return xxx_messageInfo_FutureMatches.Size(m)
-}
-func (m *FutureMatches) XXX_DiscardUnknown() {
-	xxx_messageInfo_FutureMatches.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_FutureMatches proto.InternalMessageInfo
-
-func (m *FutureMatches) GetMatches() map[string]string {
-	if m != nil {
-		return m.Matches
+func (x *FutureMatches) GetMatches() map[string]string {
+	if x != nil {
+		return x.Matches
 	}
 	return nil
 }
 
-func init() {
-	proto.RegisterEnum("model.TeamMatchResult_Result", TeamMatchResult_Result_name, TeamMatchResult_Result_value)
-	proto.RegisterEnum("model.TeamMatchResult_Location", TeamMatchResult_Location_name, TeamMatchResult_Location_value)
-	proto.RegisterType((*Matches)(nil), "model.Matches")
-	proto.RegisterType((*Match)(nil), "model.Match")
-	proto.RegisterType((*Forms)(nil), "model.Forms")
-	proto.RegisterMapType((map[string]*TeamForm)(nil), "model.Forms.TeamsEntry")
-	proto.RegisterType((*TeamForm)(nil), "model.TeamForm")
-	proto.RegisterType((*TeamMatchResult)(nil), "model.TeamMatchResult")
-	proto.RegisterType((*FutureMatches)(nil), "model.FutureMatches")
-	proto.RegisterMapType((map[string]string)(nil), "model.FutureMatches.MatchesEntry")
+var File_fixture_proto protoreflect.FileDescriptor
+
+var file_fixture_proto_rawDesc = []byte{
+	0x0a, 0x0d, 0x66, 0x69, 0x78, 0x74, 0x75, 0x72, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
+	0x05, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70,
+	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74,
+	0x6f, 0x62, 0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70,
+	0x72, 0x6f, 0x74, 0x6f, 0x22, 0x31, 0x0a, 0x07, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x12,
+	0x26, 0x0a, 0x07, 0x6d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b,
+	0x32, 0x0c, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x52, 0x07,
+	0x6d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x22, 0xdf, 0x01, 0x0a, 0x05, 0x4d, 0x61, 0x74, 0x63,
+	0x68, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69,
+	0x64, 0x12, 0x14, 0x0a, 0x05, 0x68, 0x54, 0x65, 0x61, 0x6d, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x05, 0x68, 0x54, 0x65, 0x61, 0x6d, 0x12, 0x14, 0x0a, 0x05, 0x61, 0x54, 0x65, 0x61, 0x6d,
+	0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x61, 0x54, 0x65, 0x61, 0x6d, 0x12, 0x16, 0x0a,
+	0x06, 0x68, 0x47, 0x6f, 0x61, 0x6c, 0x73, 0x18, 0x04, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x68,
+	0x47, 0x6f, 0x61, 0x6c, 0x73, 0x12, 0x16, 0x0a, 0x06, 0x61, 0x47, 0x6f, 0x61, 0x6c, 0x73, 0x18,
+	0x05, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x61, 0x47, 0x6f, 0x61, 0x6c, 0x73, 0x12, 0x16, 0x0a,
+	0x06, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x64, 0x18, 0x06, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x70,
+	0x6c, 0x61, 0x79, 0x65, 0x64, 0x12, 0x36, 0x0a, 0x08, 0x64, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d,
+	0x65, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
+	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74,
+	0x61, 0x6d, 0x70, 0x52, 0x08, 0x64, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x1a, 0x0a,
+	0x08, 0x6d, 0x61, 0x74, 0x63, 0x68, 0x64, 0x61, 0x79, 0x18, 0x08, 0x20, 0x01, 0x28, 0x05, 0x52,
+	0x08, 0x6d, 0x61, 0x74, 0x63, 0x68, 0x64, 0x61, 0x79, 0x22, 0x81, 0x01, 0x0a, 0x05, 0x46, 0x6f,
+	0x72, 0x6d, 0x73, 0x12, 0x2d, 0x0a, 0x05, 0x74, 0x65, 0x61, 0x6d, 0x73, 0x18, 0x01, 0x20, 0x03,
+	0x28, 0x0b, 0x32, 0x17, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x46, 0x6f, 0x72, 0x6d, 0x73,
+	0x2e, 0x54, 0x65, 0x61, 0x6d, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05, 0x74, 0x65, 0x61,
+	0x6d, 0x73, 0x1a, 0x49, 0x0a, 0x0a, 0x54, 0x65, 0x61, 0x6d, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79,
+	0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b,
+	0x65, 0x79, 0x12, 0x25, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x0b, 0x32, 0x0f, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x54, 0x65, 0x61, 0x6d, 0x46, 0x6f,
+	0x72, 0x6d, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x22, 0x38, 0x0a,
+	0x08, 0x54, 0x65, 0x61, 0x6d, 0x46, 0x6f, 0x72, 0x6d, 0x12, 0x2c, 0x0a, 0x05, 0x66, 0x6f, 0x72,
+	0x6d, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c,
+	0x2e, 0x54, 0x65, 0x61, 0x6d, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74,
+	0x52, 0x05, 0x66, 0x6f, 0x72, 0x6d, 0x73, 0x22, 0xfe, 0x01, 0x0a, 0x0f, 0x54, 0x65, 0x61, 0x6d,
+	0x4d, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x12, 0x35, 0x0a, 0x06, 0x72,
+	0x65, 0x73, 0x75, 0x6c, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x1d, 0x2e, 0x6d, 0x6f,
+	0x64, 0x65, 0x6c, 0x2e, 0x54, 0x65, 0x61, 0x6d, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73,
+	0x75, 0x6c, 0x74, 0x2e, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x06, 0x72, 0x65, 0x73, 0x75,
+	0x6c, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x6f, 0x70, 0x70, 0x6f,
+	0x6e, 0x65, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x6f, 0x70, 0x70, 0x6f,
+	0x6e, 0x65, 0x6e, 0x74, 0x12, 0x3b, 0x0a, 0x08, 0x6c, 0x6f, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e,
+	0x18, 0x04, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x1f, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x54,
+	0x65, 0x61, 0x6d, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x2e, 0x4c,
+	0x6f, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x08, 0x6c, 0x6f, 0x63, 0x61, 0x74, 0x69, 0x6f,
+	0x6e, 0x22, 0x25, 0x0a, 0x06, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x12, 0x07, 0x0a, 0x03, 0x57,
+	0x49, 0x4e, 0x10, 0x00, 0x12, 0x08, 0x0a, 0x04, 0x44, 0x52, 0x41, 0x57, 0x10, 0x01, 0x12, 0x08,
+	0x0a, 0x04, 0x4c, 0x4f, 0x53, 0x53, 0x10, 0x02, 0x22, 0x1e, 0x0a, 0x08, 0x4c, 0x6f, 0x63, 0x61,
+	0x74, 0x69, 0x6f, 0x6e, 0x12, 0x08, 0x0a, 0x04, 0x48, 0x4f, 0x4d, 0x45, 0x10, 0x00, 0x12, 0x08,
+	0x0a, 0x04, 0x41, 0x57, 0x41, 0x59, 0x10, 0x01, 0x22, 0x88, 0x01, 0x0a, 0x0d, 0x46, 0x75, 0x74,
+	0x75, 0x72, 0x65, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x12, 0x3b, 0x0a, 0x07, 0x6d, 0x61,
+	0x74, 0x63, 0x68, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x21, 0x2e, 0x6d, 0x6f,
+	0x64, 0x65, 0x6c, 0x2e, 0x46, 0x75, 0x74, 0x75, 0x72, 0x65, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x65,
+	0x73, 0x2e, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x07,
+	0x6d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x1a, 0x3a, 0x0a, 0x0c, 0x4d, 0x61, 0x74, 0x63, 0x68,
+	0x65, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c,
+	0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a,
+	0x02, 0x38, 0x01, 0x32, 0xc4, 0x01, 0x0a, 0x0e, 0x46, 0x69, 0x78, 0x74, 0x75, 0x72, 0x65, 0x53,
+	0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x36, 0x0a, 0x0a, 0x47, 0x65, 0x74, 0x4d, 0x61, 0x74,
+	0x63, 0x68, 0x65, 0x73, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x0e, 0x2e, 0x6d,
+	0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x22, 0x00, 0x12, 0x35,
+	0x0a, 0x0b, 0x47, 0x65, 0x74, 0x54, 0x65, 0x61, 0x6d, 0x46, 0x6f, 0x72, 0x6d, 0x12, 0x16, 0x2e,
+	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
+	0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x0c, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x46, 0x6f,
+	0x72, 0x6d, 0x73, 0x22, 0x00, 0x12, 0x43, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x46, 0x75, 0x74, 0x75,
+	0x72, 0x65, 0x46, 0x69, 0x78, 0x74, 0x75, 0x72, 0x65, 0x73, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f,
+	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70,
+	0x74, 0x79, 0x1a, 0x14, 0x2e, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x2e, 0x46, 0x75, 0x74, 0x75, 0x72,
+	0x65, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x22, 0x00, 0x42, 0x27, 0x0a, 0x23, 0x63, 0x6f,
+	0x6d, 0x2e, 0x63, 0x73, 0x68, 0x65, 0x70, 0x34, 0x2e, 0x70, 0x72, 0x65, 0x6d, 0x69, 0x65, 0x72,
+	0x70, 0x72, 0x65, 0x64, 0x69, 0x63, 0x74, 0x6f, 0x72, 0x2e, 0x66, 0x69, 0x78, 0x74, 0x75, 0x72,
+	0x65, 0x50, 0x01, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
-func init() { proto.RegisterFile("fixture.proto", fileDescriptor_d81820e403a865f4) }
+var (
+	file_fixture_proto_rawDescOnce sync.Once
+	file_fixture_proto_rawDescData = file_fixture_proto_rawDesc
+)
 
-var fileDescriptor_d81820e403a865f4 = []byte{
-	// 604 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x74, 0x54, 0xcf, 0x4e, 0xdb, 0x4e,
-	0x10, 0xce, 0x26, 0x38, 0x31, 0x13, 0x08, 0xf9, 0xad, 0x10, 0x3f, 0xcb, 0x55, 0x4b, 0xea, 0x8a,
-	0x36, 0x87, 0xd6, 0xa8, 0x69, 0x41, 0x08, 0x4e, 0xb4, 0x85, 0x14, 0x09, 0x4a, 0x65, 0x90, 0x50,
-	0x8f, 0x8b, 0x3d, 0x10, 0xab, 0x76, 0xd6, 0xb2, 0x37, 0xa8, 0x39, 0xf6, 0xd6, 0x07, 0xeb, 0x7b,
-	0xf4, 0x4d, 0xaa, 0x6a, 0xff, 0xd8, 0xb5, 0x28, 0x39, 0x79, 0xe7, 0x9b, 0x6f, 0x66, 0x67, 0xe6,
-	0x9b, 0x35, 0xac, 0xde, 0xc4, 0xdf, 0xc4, 0x2c, 0x47, 0x3f, 0xcb, 0xb9, 0xe0, 0xd4, 0x4a, 0x79,
-	0x84, 0x89, 0xfb, 0xe8, 0x96, 0xf3, 0xdb, 0x04, 0xb7, 0x15, 0x78, 0x3d, 0xbb, 0xd9, 0xc6, 0x34,
-	0x13, 0x73, 0xcd, 0x71, 0x37, 0xef, 0x3b, 0x45, 0x9c, 0x62, 0x21, 0x58, 0x9a, 0x69, 0x82, 0xf7,
-	0x1a, 0x3a, 0x67, 0x4c, 0x84, 0x13, 0x2c, 0xe8, 0x73, 0xe8, 0xa4, 0xfa, 0xe8, 0x90, 0x41, 0x6b,
-	0xd8, 0x1d, 0xad, 0xf8, 0xea, 0x06, 0x5f, 0x11, 0x82, 0xd2, 0xe9, 0xfd, 0x22, 0x60, 0x29, 0x88,
-	0xf6, 0xa0, 0x19, 0x47, 0x0e, 0x19, 0x90, 0xe1, 0x72, 0xd0, 0x8c, 0x23, 0xba, 0x0e, 0xd6, 0xe4,
-	0x12, 0x59, 0xea, 0x34, 0x15, 0xa4, 0x0d, 0x89, 0x32, 0x85, 0xb6, 0x34, 0xaa, 0x0c, 0xba, 0x01,
-	0xed, 0xc9, 0x98, 0xb3, 0xa4, 0x70, 0x96, 0x06, 0x64, 0x68, 0x05, 0xc6, 0x92, 0x38, 0xd3, 0xb8,
-	0xa5, 0x71, 0x56, 0xe1, 0x59, 0xc2, 0xe6, 0x18, 0x39, 0x6d, 0x8d, 0x6b, 0x8b, 0xee, 0x82, 0x1d,
-	0x31, 0x81, 0x97, 0x71, 0x8a, 0x4e, 0x67, 0x40, 0x86, 0xdd, 0x91, 0xeb, 0xeb, 0xa6, 0xfd, 0xb2,
-	0x69, 0xff, 0xb2, 0x6c, 0x3a, 0xa8, 0xb8, 0xd4, 0x05, 0x5b, 0x35, 0x14, 0xb1, 0xb9, 0x63, 0xab,
-	0x8c, 0x95, 0xed, 0x7d, 0x27, 0x60, 0x1d, 0xf3, 0x3c, 0x2d, 0xe8, 0x2b, 0xb0, 0x04, 0xb2, 0xb4,
-	0x9c, 0xc8, 0xff, 0x66, 0x22, 0xca, 0xe9, 0xcb, 0x3e, 0x8a, 0xa3, 0xa9, 0xc8, 0xe7, 0x81, 0x66,
-	0xb9, 0x27, 0x00, 0x7f, 0x41, 0xda, 0x87, 0xd6, 0x57, 0x9c, 0x9b, 0xf9, 0xc8, 0x23, 0xdd, 0x02,
-	0xeb, 0x8e, 0x25, 0x33, 0x54, 0x03, 0xea, 0x8e, 0xd6, 0x4c, 0x3a, 0x19, 0x23, 0x53, 0x06, 0xda,
-	0xbb, 0xdf, 0xdc, 0x23, 0xde, 0x1e, 0xd8, 0x25, 0x4c, 0x5f, 0x82, 0x75, 0x23, 0x6f, 0x34, 0x55,
-	0x6c, 0xd4, 0xc2, 0xb4, 0x36, 0x58, 0xcc, 0x12, 0x11, 0x68, 0x92, 0xf7, 0x9b, 0xc0, 0xda, 0x3d,
-	0x17, 0xdd, 0x81, 0x76, 0xae, 0x4e, 0xaa, 0x9a, 0xde, 0xe8, 0xf1, 0xc3, 0x29, 0x7c, 0x93, 0xc9,
-	0x90, 0xa5, 0x74, 0x45, 0xc8, 0x73, 0x2c, 0x05, 0x55, 0x86, 0x1c, 0x1d, 0xcf, 0x32, 0x3e, 0xc5,
-	0xa9, 0x30, 0x9a, 0x56, 0x36, 0x3d, 0x00, 0x3b, 0xe1, 0x21, 0x13, 0x31, 0x9f, 0x2a, 0x61, 0x7b,
-	0xa3, 0xcd, 0x05, 0x57, 0x9d, 0x1a, 0x5a, 0x50, 0x05, 0x78, 0x5b, 0xd0, 0x36, 0xf5, 0x76, 0xa0,
-	0x75, 0x75, 0xf2, 0xa9, 0xdf, 0xa0, 0x36, 0x2c, 0x7d, 0x08, 0x0e, 0xaf, 0xfa, 0x44, 0x9e, 0x4e,
-	0xcf, 0x2f, 0x2e, 0xfa, 0x4d, 0xef, 0x09, 0xd8, 0x65, 0xb0, 0x44, 0x3f, 0x9e, 0x9f, 0x1d, 0x69,
-	0xe6, 0xe1, 0xd5, 0xe1, 0x97, 0x3e, 0xf1, 0x7e, 0x10, 0x58, 0x3d, 0x9e, 0xc9, 0x97, 0x52, 0xae,
-	0xf6, 0xc1, 0xfd, 0xd5, 0x7e, 0x5a, 0x0a, 0x59, 0xa7, 0xf9, 0xe6, 0xab, 0x25, 0x2d, 0x23, 0xdc,
-	0x7d, 0x58, 0xa9, 0x3b, 0x1e, 0x90, 0x75, 0xbd, 0x2e, 0xeb, 0x72, 0x4d, 0xc5, 0xd1, 0x4f, 0x02,
-	0xbd, 0x63, 0xfd, 0x6a, 0x2f, 0x30, 0xbf, 0x8b, 0x43, 0xa4, 0xbb, 0x00, 0x63, 0x14, 0x65, 0x65,
-	0x1b, 0xff, 0x2c, 0xeb, 0x91, 0x7c, 0xbe, 0x6e, 0xaf, 0xfe, 0xf6, 0xb0, 0xf0, 0x1a, 0x74, 0x07,
-	0xba, 0x63, 0x14, 0xd5, 0x4e, 0x2c, 0x0a, 0x5c, 0xa9, 0xaf, 0xa8, 0xd7, 0xa0, 0xef, 0xe1, 0xbf,
-	0x31, 0x0a, 0xdd, 0xa7, 0xa9, 0x64, 0xf1, 0xad, 0xeb, 0x0f, 0x8d, 0xc5, 0x6b, 0xbc, 0x7b, 0x01,
-	0xcf, 0x42, 0x9e, 0xfa, 0x61, 0x31, 0xc1, 0xec, 0xad, 0x9f, 0xe5, 0x98, 0xc6, 0x98, 0x67, 0x39,
-	0x46, 0x71, 0x28, 0x78, 0xee, 0x9b, 0xff, 0xd2, 0x67, 0x72, 0xdd, 0x56, 0x09, 0xdf, 0xfc, 0x09,
-	0x00, 0x00, 0xff, 0xff, 0x6d, 0x1a, 0x8b, 0xfb, 0xab, 0x04, 0x00, 0x00,
+func file_fixture_proto_rawDescGZIP() []byte {
+	file_fixture_proto_rawDescOnce.Do(func() {
+		file_fixture_proto_rawDescData = protoimpl.X.CompressGZIP(file_fixture_proto_rawDescData)
+	})
+	return file_fixture_proto_rawDescData
+}
+
+var file_fixture_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
+var file_fixture_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
+var file_fixture_proto_goTypes = []interface{}{
+	(TeamMatchResult_Result)(0),   // 0: model.TeamMatchResult.Result
+	(TeamMatchResult_Location)(0), // 1: model.TeamMatchResult.Location
+	(*Matches)(nil),               // 2: model.Matches
+	(*Match)(nil),                 // 3: model.Match
+	(*Forms)(nil),                 // 4: model.Forms
+	(*TeamForm)(nil),              // 5: model.TeamForm
+	(*TeamMatchResult)(nil),       // 6: model.TeamMatchResult
+	(*FutureMatches)(nil),         // 7: model.FutureMatches
+	nil,                           // 8: model.Forms.TeamsEntry
+	nil,                           // 9: model.FutureMatches.MatchesEntry
+	(*timestamp.Timestamp)(nil),   // 10: google.protobuf.Timestamp
+	(*empty.Empty)(nil),           // 11: google.protobuf.Empty
+}
+var file_fixture_proto_depIdxs = []int32{
+	3,  // 0: model.Matches.matches:type_name -> model.Match
+	10, // 1: model.Match.dateTime:type_name -> google.protobuf.Timestamp
+	8,  // 2: model.Forms.teams:type_name -> model.Forms.TeamsEntry
+	6,  // 3: model.TeamForm.forms:type_name -> model.TeamMatchResult
+	0,  // 4: model.TeamMatchResult.result:type_name -> model.TeamMatchResult.Result
+	1,  // 5: model.TeamMatchResult.location:type_name -> model.TeamMatchResult.Location
+	9,  // 6: model.FutureMatches.matches:type_name -> model.FutureMatches.MatchesEntry
+	5,  // 7: model.Forms.TeamsEntry.value:type_name -> model.TeamForm
+	11, // 8: model.FixtureService.GetMatches:input_type -> google.protobuf.Empty
+	11, // 9: model.FixtureService.GetTeamForm:input_type -> google.protobuf.Empty
+	11, // 10: model.FixtureService.GetFutureFixtures:input_type -> google.protobuf.Empty
+	2,  // 11: model.FixtureService.GetMatches:output_type -> model.Matches
+	4,  // 12: model.FixtureService.GetTeamForm:output_type -> model.Forms
+	7,  // 13: model.FixtureService.GetFutureFixtures:output_type -> model.FutureMatches
+	11, // [11:14] is the sub-list for method output_type
+	8,  // [8:11] is the sub-list for method input_type
+	8,  // [8:8] is the sub-list for extension type_name
+	8,  // [8:8] is the sub-list for extension extendee
+	0,  // [0:8] is the sub-list for field type_name
+}
+
+func init() { file_fixture_proto_init() }
+func file_fixture_proto_init() {
+	if File_fixture_proto != nil {
+		return
+	}
+	if !protoimpl.UnsafeEnabled {
+		file_fixture_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Matches); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fixture_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Match); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fixture_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Forms); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fixture_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TeamForm); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fixture_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TeamMatchResult); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fixture_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*FutureMatches); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+	}
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: file_fixture_proto_rawDesc,
+			NumEnums:      2,
+			NumMessages:   8,
+			NumExtensions: 0,
+			NumServices:   1,
+		},
+		GoTypes:           file_fixture_proto_goTypes,
+		DependencyIndexes: file_fixture_proto_depIdxs,
+		EnumInfos:         file_fixture_proto_enumTypes,
+		MessageInfos:      file_fixture_proto_msgTypes,
+	}.Build()
+	File_fixture_proto = out.File
+	file_fixture_proto_rawDesc = nil
+	file_fixture_proto_goTypes = nil
+	file_fixture_proto_depIdxs = nil
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ context.Context
-var _ grpc.ClientConn
+var _ grpc.ClientConnInterface
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion4
+const _ = grpc.SupportPackageIsVersion6
 
 // FixtureServiceClient is the client API for FixtureService service.
 //
@@ -468,10 +736,10 @@ type FixtureServiceClient interface {
 }
 
 type fixtureServiceClient struct {
-	cc *grpc.ClientConn
+	cc grpc.ClientConnInterface
 }
 
-func NewFixtureServiceClient(cc *grpc.ClientConn) FixtureServiceClient {
+func NewFixtureServiceClient(cc grpc.ClientConnInterface) FixtureServiceClient {
 	return &fixtureServiceClient{cc}
 }
 
@@ -513,13 +781,13 @@ type FixtureServiceServer interface {
 type UnimplementedFixtureServiceServer struct {
 }
 
-func (*UnimplementedFixtureServiceServer) GetMatches(ctx context.Context, req *empty.Empty) (*Matches, error) {
+func (*UnimplementedFixtureServiceServer) GetMatches(context.Context, *empty.Empty) (*Matches, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetMatches not implemented")
 }
-func (*UnimplementedFixtureServiceServer) GetTeamForm(ctx context.Context, req *empty.Empty) (*Forms, error) {
+func (*UnimplementedFixtureServiceServer) GetTeamForm(context.Context, *empty.Empty) (*Forms, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetTeamForm not implemented")
 }
-func (*UnimplementedFixtureServiceServer) GetFutureFixtures(ctx context.Context, req *empty.Empty) (*FutureMatches, error) {
+func (*UnimplementedFixtureServiceServer) GetFutureFixtures(context.Context, *empty.Empty) (*FutureMatches, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetFutureFixtures not implemented")
 }
 
