@@ -1,164 +1,283 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.25.0
+// 	protoc        v3.12.3
 // source: request.proto
 
 package model
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
+import (
+	proto "github.com/golang/protobuf/proto"
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
+)
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+// This is a compile-time assertion that a sufficiently up-to-date version
+// of the legacy proto package is being used.
+const _ = proto.ProtoPackageIsVersion4
 
 type IdRequest struct {
-	Id                   string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Id string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 }
 
-func (m *IdRequest) Reset()         { *m = IdRequest{} }
-func (m *IdRequest) String() string { return proto.CompactTextString(m) }
-func (*IdRequest) ProtoMessage()    {}
+func (x *IdRequest) Reset() {
+	*x = IdRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_request_proto_msgTypes[0]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *IdRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*IdRequest) ProtoMessage() {}
+
+func (x *IdRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_request_proto_msgTypes[0]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use IdRequest.ProtoReflect.Descriptor instead.
 func (*IdRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_request_ff668dceef9c813d, []int{0}
-}
-func (m *IdRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_IdRequest.Unmarshal(m, b)
-}
-func (m *IdRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_IdRequest.Marshal(b, m, deterministic)
-}
-func (dst *IdRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_IdRequest.Merge(dst, src)
-}
-func (m *IdRequest) XXX_Size() int {
-	return xxx_messageInfo_IdRequest.Size(m)
-}
-func (m *IdRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_IdRequest.DiscardUnknown(m)
+	return file_request_proto_rawDescGZIP(), []int{0}
 }
 
-var xxx_messageInfo_IdRequest proto.InternalMessageInfo
-
-func (m *IdRequest) GetId() string {
-	if m != nil {
-		return m.Id
+func (x *IdRequest) GetId() string {
+	if x != nil {
+		return x.Id
 	}
 	return ""
 }
 
 type EmailRequest struct {
-	Email                string   `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Email string `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
 }
 
-func (m *EmailRequest) Reset()         { *m = EmailRequest{} }
-func (m *EmailRequest) String() string { return proto.CompactTextString(m) }
-func (*EmailRequest) ProtoMessage()    {}
+func (x *EmailRequest) Reset() {
+	*x = EmailRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_request_proto_msgTypes[1]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *EmailRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*EmailRequest) ProtoMessage() {}
+
+func (x *EmailRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_request_proto_msgTypes[1]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use EmailRequest.ProtoReflect.Descriptor instead.
 func (*EmailRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_request_ff668dceef9c813d, []int{1}
-}
-func (m *EmailRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_EmailRequest.Unmarshal(m, b)
-}
-func (m *EmailRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_EmailRequest.Marshal(b, m, deterministic)
-}
-func (dst *EmailRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_EmailRequest.Merge(dst, src)
-}
-func (m *EmailRequest) XXX_Size() int {
-	return xxx_messageInfo_EmailRequest.Size(m)
-}
-func (m *EmailRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_EmailRequest.DiscardUnknown(m)
+	return file_request_proto_rawDescGZIP(), []int{1}
 }
 
-var xxx_messageInfo_EmailRequest proto.InternalMessageInfo
-
-func (m *EmailRequest) GetEmail() string {
-	if m != nil {
-		return m.Email
+func (x *EmailRequest) GetEmail() string {
+	if x != nil {
+		return x.Email
 	}
 	return ""
 }
 
 type PredictionRequest struct {
-	UserId               string   `protobuf:"bytes,1,opt,name=userId,proto3" json:"userId,omitempty"`
-	MatchId              string   `protobuf:"bytes,2,opt,name=matchId,proto3" json:"matchId,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	UserId  string `protobuf:"bytes,1,opt,name=userId,proto3" json:"userId,omitempty"`
+	MatchId string `protobuf:"bytes,2,opt,name=matchId,proto3" json:"matchId,omitempty"`
 }
 
-func (m *PredictionRequest) Reset()         { *m = PredictionRequest{} }
-func (m *PredictionRequest) String() string { return proto.CompactTextString(m) }
-func (*PredictionRequest) ProtoMessage()    {}
+func (x *PredictionRequest) Reset() {
+	*x = PredictionRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_request_proto_msgTypes[2]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *PredictionRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*PredictionRequest) ProtoMessage() {}
+
+func (x *PredictionRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_request_proto_msgTypes[2]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use PredictionRequest.ProtoReflect.Descriptor instead.
 func (*PredictionRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_request_ff668dceef9c813d, []int{2}
-}
-func (m *PredictionRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_PredictionRequest.Unmarshal(m, b)
-}
-func (m *PredictionRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_PredictionRequest.Marshal(b, m, deterministic)
-}
-func (dst *PredictionRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_PredictionRequest.Merge(dst, src)
-}
-func (m *PredictionRequest) XXX_Size() int {
-	return xxx_messageInfo_PredictionRequest.Size(m)
-}
-func (m *PredictionRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_PredictionRequest.DiscardUnknown(m)
+	return file_request_proto_rawDescGZIP(), []int{2}
 }
 
-var xxx_messageInfo_PredictionRequest proto.InternalMessageInfo
-
-func (m *PredictionRequest) GetUserId() string {
-	if m != nil {
-		return m.UserId
+func (x *PredictionRequest) GetUserId() string {
+	if x != nil {
+		return x.UserId
 	}
 	return ""
 }
 
-func (m *PredictionRequest) GetMatchId() string {
-	if m != nil {
-		return m.MatchId
+func (x *PredictionRequest) GetMatchId() string {
+	if x != nil {
+		return x.MatchId
 	}
 	return ""
 }
 
-func init() {
-	proto.RegisterType((*IdRequest)(nil), "model.IdRequest")
-	proto.RegisterType((*EmailRequest)(nil), "model.EmailRequest")
-	proto.RegisterType((*PredictionRequest)(nil), "model.PredictionRequest")
+var File_request_proto protoreflect.FileDescriptor
+
+var file_request_proto_rawDesc = []byte{
+	0x0a, 0x0d, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
+	0x05, 0x6d, 0x6f, 0x64, 0x65, 0x6c, 0x22, 0x1b, 0x0a, 0x09, 0x49, 0x64, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x02, 0x69, 0x64, 0x22, 0x24, 0x0a, 0x0c, 0x45, 0x6d, 0x61, 0x69, 0x6c, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x18, 0x01, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x22, 0x45, 0x0a, 0x11, 0x50, 0x72, 0x65,
+	0x64, 0x69, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x16,
+	0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06,
+	0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x61, 0x74, 0x63, 0x68, 0x49,
+	0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x61, 0x74, 0x63, 0x68, 0x49, 0x64,
+	0x42, 0x27, 0x0a, 0x23, 0x63, 0x6f, 0x6d, 0x2e, 0x63, 0x73, 0x68, 0x65, 0x70, 0x34, 0x2e, 0x70,
+	0x72, 0x65, 0x6d, 0x69, 0x65, 0x72, 0x70, 0x72, 0x65, 0x64, 0x69, 0x63, 0x74, 0x6f, 0x72, 0x2e,
+	0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x50, 0x01, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x33,
 }
 
-func init() { proto.RegisterFile("request.proto", fileDescriptor_request_ff668dceef9c813d) }
+var (
+	file_request_proto_rawDescOnce sync.Once
+	file_request_proto_rawDescData = file_request_proto_rawDesc
+)
 
-var fileDescriptor_request_ff668dceef9c813d = []byte{
-	// 172 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x34, 0x8f, 0xc1, 0xea, 0x82, 0x40,
-	0x10, 0x87, 0x51, 0xd0, 0x3f, 0x0e, 0xff, 0x82, 0x24, 0x42, 0xe8, 0x12, 0x16, 0xd4, 0x69, 0x2f,
-	0xf5, 0x04, 0x81, 0x07, 0x6f, 0xe2, 0x1b, 0xd8, 0xee, 0x80, 0x03, 0x6e, 0xb3, 0x8d, 0xeb, 0xfb,
-	0x47, 0xea, 0x1e, 0x3f, 0xbe, 0x1f, 0x1f, 0x33, 0xb0, 0x11, 0xfc, 0x4c, 0x38, 0x7a, 0xe5, 0x84,
-	0x3d, 0xe7, 0x89, 0x65, 0x83, 0x43, 0x79, 0x84, 0xac, 0x36, 0xed, 0x62, 0xf2, 0x2d, 0xc4, 0x64,
-	0x8a, 0xe8, 0x14, 0xdd, 0xb2, 0x36, 0x26, 0x53, 0x5e, 0xe0, 0xbf, 0xb2, 0x1d, 0x0d, 0xc1, 0xef,
-	0x21, 0xc1, 0x1f, 0xaf, 0x93, 0x05, 0xca, 0x0a, 0x76, 0x8d, 0xa0, 0x21, 0xed, 0x89, 0xdf, 0x61,
-	0x7a, 0x80, 0x74, 0x1a, 0x51, 0xea, 0x90, 0x5b, 0x29, 0x2f, 0xe0, 0xcf, 0x76, 0x5e, 0xf7, 0xb5,
-	0x29, 0xe2, 0x59, 0x04, 0x7c, 0x5e, 0xe1, 0xac, 0xd9, 0x2a, 0x3d, 0xf6, 0xe8, 0x1e, 0xca, 0x09,
-	0x5a, 0x42, 0x71, 0x4b, 0x98, 0x45, 0xad, 0xd7, 0x37, 0xd1, 0x2b, 0x9d, 0x1f, 0xb8, 0x7f, 0x03,
-	0x00, 0x00, 0xff, 0xff, 0x2c, 0xe5, 0x6f, 0x3c, 0xd1, 0x00, 0x00, 0x00,
+func file_request_proto_rawDescGZIP() []byte {
+	file_request_proto_rawDescOnce.Do(func() {
+		file_request_proto_rawDescData = protoimpl.X.CompressGZIP(file_request_proto_rawDescData)
+	})
+	return file_request_proto_rawDescData
+}
+
+var file_request_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
+var file_request_proto_goTypes = []interface{}{
+	(*IdRequest)(nil),         // 0: model.IdRequest
+	(*EmailRequest)(nil),      // 1: model.EmailRequest
+	(*PredictionRequest)(nil), // 2: model.PredictionRequest
+}
+var file_request_proto_depIdxs = []int32{
+	0, // [0:0] is the sub-list for method output_type
+	0, // [0:0] is the sub-list for method input_type
+	0, // [0:0] is the sub-list for extension type_name
+	0, // [0:0] is the sub-list for extension extendee
+	0, // [0:0] is the sub-list for field type_name
+}
+
+func init() { file_request_proto_init() }
+func file_request_proto_init() {
+	if File_request_proto != nil {
+		return
+	}
+	if !protoimpl.UnsafeEnabled {
+		file_request_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*IdRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_request_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*EmailRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_request_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*PredictionRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+	}
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: file_request_proto_rawDesc,
+			NumEnums:      0,
+			NumMessages:   3,
+			NumExtensions: 0,
+			NumServices:   0,
+		},
+		GoTypes:           file_request_proto_goTypes,
+		DependencyIndexes: file_request_proto_depIdxs,
+		MessageInfos:      file_request_proto_msgTypes,
+	}.Build()
+	File_request_proto = out.File
+	file_request_proto_rawDesc = nil
+	file_request_proto_goTypes = nil
+	file_request_proto_depIdxs = nil
 }
