@@ -1,39 +1,82 @@
 import * as grpc from "grpc";
 import {MatchFacts, matchFactsFromGrpc} from "../../model/live/matchFacts";
+import {SubscriptionContext, TokenContext, tokenMetadata} from "../../model/context/context";
+import {
+    GetMatchRequest,
+    GetMatchResponse, ListTodaysMatchesResponse,
+    LiveMatchClient, MatchSummaryResponse,
+    UpcomingMatchesGRPCResult,
+    UpcomingMatchesResult
+} from "./client";
+import {logger} from "../../log/logger";
+import {secondsToDateString} from "../../util/DateUtils";
+import {PubSub} from "graphql-subscriptions";
+import WebSocket from "ws";
+import {ClientReadableStream} from "grpc";
+import {handleSubscription, withCancel} from "../../util/subsciptionUtils";
 
-interface Params {
-    token: string;
-}
-
-class UpcomingMatchesGRPCResult {
-    matches: Map<string, any> = new Map<string, any>();
-}
-
-class UpcomingMatchesResult {
-    matches: Matches[] = [];
-}
-
-class Matches {
-    date!: string;
-    matches: MatchFacts[] = [];
-}
 
 export class LiveMatch {
-    constructor(private client: any, private pubsub: any) {
+    constructor(private client: LiveMatchClient, private pubsub: PubSub) {
     }
 
-    public getUpcomingMatches(_obj: any, args: any, context: any) {
-        const event = "upcoming";
-        this.getUpcoming(context, event);
-        return this.pubsub.asyncIterator([event]);
-    };
+    public getMatch(ctx: TokenContext, req: GetMatchRequest) {
+        return new Promise((resolve: any, reject: any) => {
+            this.client.getLiveMatch(req, tokenMetadata(ctx), (err: grpc.ServiceError, res: GetMatchResponse) => {
+                if (err) {
+                    logger.error({
+                        "message": "get_live_match_error",
+                        "error": {
+                            "code": err.code,
+                            "details": err.details,
+                            "message": err.message,
+                        },
+                        "matchId": req.id,
+                    });
+                    return reject(err.message);
+                }
 
-    private getUpcoming(context: any, event: string) {
-        const pubsub = this.pubsub;
-        const metadata = new grpc.Metadata();
-        metadata.add('token', context.token);
-        const call = this.client.getUpcomingMatches({}, metadata);
-        call.on('data', function (fixtures: UpcomingMatchesGRPCResult) {
+                res.match.matchDate = secondsToDateString(res.match.matchDate.seconds);
+
+                resolve(res.match);
+            });
+        });
+    }
+
+    public listTodaysMatches(ctx: TokenContext) {
+        return new Promise((resolve: any, reject: any) => {
+            this.client.listTodaysMatches({}, tokenMetadata(ctx), (err: grpc.ServiceError, res: ListTodaysMatchesResponse) => {
+                if (err) {
+                    logger.error({
+                        "message": "list_todays_matches_error",
+                        "error": {
+                            "code": err.code,
+                            "details": err.details,
+                            "message": err.message,
+                        },
+                    });
+                    return reject(err.message);
+                }
+
+                if (!res.matches) {
+                    return resolve([]);
+                }
+
+                resolve(res.matches.map(m => {
+                    m.matchDate = secondsToDateString(m.matchDate.seconds);
+                    return m;
+                }));
+            });
+        });
+    }
+
+    public getUpcomingMatches(_obj: any, args: any, ctx: SubscriptionContext) {
+        const event = "upcoming";
+        const pubsub = new PubSub();
+
+        const call = this.client.getUpcomingMatches({}, tokenMetadata(ctx));
+
+        handleSubscription(ctx.webSocket, call, function (fixtures: UpcomingMatchesGRPCResult) {
             fixtures.matches = new Map(Object.entries(fixtures.matches));
             const result: UpcomingMatchesResult = new UpcomingMatchesResult();
             fixtures.matches.forEach((value: any, key: any) => {
@@ -48,17 +91,61 @@ export class LiveMatch {
                 upcomingMatches: result
             });
         });
-        call.on('end', function () {
-            // The server has finished sending
-            console.log("end");
+
+        return withCancel(pubsub.asyncIterator([event]), () => {
+            call.cancel();
+            console.log(`Subscription closed, do your cleanup`);
         });
-        call.on('error', function (err: any) {
-            // An error has occurred and the stream has been closed.
-            console.log("error: " + err);
+    };
+
+    public getMatchSummary(_obj: any, req: any, ctx: SubscriptionContext) {
+        const event = req.request.matchId;
+        const pubsub = new PubSub();
+
+        const call = this.client.getMatchSummary(req.request, tokenMetadata(ctx));
+
+        handleSubscription(ctx.webSocket, call, function (res: MatchSummaryResponse) {
+            pubsub.publish(event, {
+                liveMatchSummary: {
+                    liveMatch: matchFactsFromGrpc(res.match),
+                    predictionSummary: {
+                        homeWin: res.predictionSummary.homeWin ? res.predictionSummary.homeWin : 0,
+                        draw: res.predictionSummary.draw ? res.predictionSummary.draw : 0,
+                        awayWin: res.predictionSummary.awayWin ? res.predictionSummary.awayWin : 0,
+                    },
+                    prediction: res.prediction ? {
+                        userId: res.prediction.userId,
+                        matchId: res.prediction.matchId,
+                        hGoals: res.prediction.hGoals ? res.prediction.hGoals : 0,
+                        aGoals: res.prediction.aGoals ? res.prediction.aGoals : 0,
+                    } : null,
+                }
+            });
         });
-        call.on('status', function (status: any) {
-            // process status
-            console.log("status: " + status);
+
+        return withCancel(pubsub.asyncIterator([event]), () => {
+            call.cancel();
+            console.log(`Subscription closed, do your cleanup`);
+        });
+    }
+
+    public getTodaysLiveMatches(_obj: any, req: any, ctx: SubscriptionContext) {
+        const event = "today";
+        const pubsub = new PubSub();
+
+        const call = this.client.getTodaysLiveMatches({}, tokenMetadata(ctx));
+
+        handleSubscription(ctx.webSocket, call, function (res: GetMatchResponse) {
+            pubsub.publish(event, {
+                todaysLiveMatches: {
+                    match: matchFactsFromGrpc(res.match),
+                }
+            });
+        });
+
+        return withCancel(pubsub.asyncIterator([event]), () => {
+            call.cancel();
+            console.log(`Subscription closed, do your cleanup`);
         });
     }
 }
