@@ -1,64 +1,82 @@
-//go:generate mockgen -destination=./mocks/mock_user_service_client.go -package=usermocks github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen UserServiceClient
-
 package user
 
 import (
 	"context"
 	"errors"
-	gen "github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen"
-	"github.com/cshep4/premier-predictor-microservices/src/leagueservice/internal/model"
-	usermocks "github.com/cshep4/premier-predictor-microservices/src/leagueservice/internal/user/mocks"
-	"github.com/golang/mock/gomock"
-	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc/metadata"
 	"strconv"
 	"testing"
+
+	"github.com/cshep4/premier-predictor-microservices/src/leagueservice/internal/mocks/token"
+	"github.com/cshep4/premier-predictor-microservices/src/leagueservice/internal/mocks/user"
+	"github.com/cshep4/premier-predictor-microservices/src/leagueservice/internal/model"
+
+	pb "github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen"
+	"github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 const (
-	id1 = "🆔"
-	id2 = "⚽️"
+	id1   = "🆔"
+	id2   = "⚽️"
+	token = "🔑"
 )
 
 func TestUserService_GetAllUsers(t *testing.T) {
-	ctrl := gomock.NewController(t)
+	ctrl, ctx := gomock.WithContext(context.Background(), t)
 	defer ctrl.Finish()
 
-	userClient := usermocks.NewMockUserServiceClient(ctrl)
+	userClient := user_mock.NewMockUserServiceClient(ctrl)
+	tokenGenerator := token_mock.NewMockTokenGenerator(ctrl)
 
-	userService, err := NewUserService(userClient)
+	service, err := New(tokenGenerator, userClient)
 	require.NoError(t, err)
 
-	var users []*gen.User
-	var expectedUsers []*model.LeagueUser
+	var users []*pb.User
+	var expectedUsers []model.LeagueUser
 
-	for i:=0; i<200000; i++ {
-		users = append(users, &gen.User{
+	for i := 0; i < 200000; i++ {
+		users = append(users, &pb.User{
 			Id: strconv.Itoa(i),
 		})
-		expectedUsers = append(expectedUsers, &model.LeagueUser{
+		expectedUsers = append(expectedUsers, model.LeagueUser{
 			Id: strconv.Itoa(i),
 		})
 	}
 
-	resp := &gen.UserResponse{
+	resp := &pb.GetAllUsersResponse{
 		Users: users,
 	}
 
-	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
-		userClient.EXPECT().GetAllUsers(context.Background(), &empty.Empty{}).Return(nil, errors.New("error"))
+	t.Run("returns error if error generating token", func(t *testing.T) {
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return("", errors.New("error"))
 
-		result, err := userService.GetAllUsers()
+		result, err := service.GetAllUsers(ctx)
+		require.Error(t, err)
+
+		assert.Empty(t, result)
+	})
+
+	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
+
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetAllUsers(expectedCtx, &pb.GetAllUsersRequest{}).Return(nil, errors.New("error"))
+
+		result, err := service.GetAllUsers(ctx)
 		require.Error(t, err)
 
 		assert.Empty(t, result)
 	})
 
 	t.Run("gets all users from UserService", func(t *testing.T) {
-		userClient.EXPECT().GetAllUsers(context.Background(), &empty.Empty{}).Return(resp, nil)
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetAllUsers()
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetAllUsers(expectedCtx, &pb.GetAllUsersRequest{}).Return(resp, nil)
+
+		result, err := service.GetAllUsers(ctx)
 		require.NoError(t, err)
 
 		assert.Equal(t, expectedUsers, result)
@@ -66,50 +84,65 @@ func TestUserService_GetAllUsers(t *testing.T) {
 }
 
 func TestUserService_GetLeagueUsers(t *testing.T) {
-	ctrl := gomock.NewController(t)
+	ctrl, ctx := gomock.WithContext(context.Background(), t)
 	defer ctrl.Finish()
 
-	userClient := usermocks.NewMockUserServiceClient(ctrl)
+	userClient := user_mock.NewMockUserServiceClient(ctrl)
+	tokenGenerator := token_mock.NewMockTokenGenerator(ctrl)
 
-	userService, err := NewUserService(userClient)
+	service, err := New(tokenGenerator, userClient)
 	require.NoError(t, err)
 
-	var users []*gen.User
-	var expectedUsers []*model.LeagueUser
+	var users []*pb.User
+	var expectedUsers []model.LeagueUser
 
-	for i:=0; i<5; i++ {
-		users = append(users, &gen.User{
+	for i := 0; i < 5; i++ {
+		users = append(users, &pb.User{
 			Id: strconv.Itoa(i),
 		})
-		expectedUsers = append(expectedUsers, &model.LeagueUser{
+		expectedUsers = append(expectedUsers, model.LeagueUser{
 			Id: strconv.Itoa(i),
 		})
 	}
 
 	ids := []string{id1, id2}
 
-	req := &gen.GroupIdRequest{
+	req := &pb.GetUsersByIdsRequest{
 		Ids: ids,
 	}
 
-	resp := &gen.UserResponse{
+	resp := &pb.GetUsersByIdsResponse{
 		Users: users,
 	}
 
+	t.Run("returns error if error generating token", func(t *testing.T) {
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return("", errors.New("error"))
+
+		result, err := service.GetLeagueUsers(ctx, ids)
+		require.Error(t, err)
+
+		assert.Empty(t, result)
+	})
 
 	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
-		userClient.EXPECT().GetAllUsersByIds(context.Background(), req).Return(nil, errors.New("error"))
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetLeagueUsers(ids)
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetUsersByIds(expectedCtx, req).Return(nil, errors.New("error"))
+
+		result, err := service.GetLeagueUsers(ctx, ids)
 		require.Error(t, err)
 
 		assert.Empty(t, result)
 	})
 
 	t.Run("gets all users from UserService", func(t *testing.T) {
-		userClient.EXPECT().GetAllUsersByIds(context.Background(), req).Return(resp, nil)
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetLeagueUsers(ids)
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetUsersByIds(expectedCtx, req).Return(resp, nil)
+
+		result, err := service.GetLeagueUsers(ctx, ids)
 		require.NoError(t, err)
 
 		assert.Equal(t, expectedUsers, result)
@@ -117,38 +150,53 @@ func TestUserService_GetLeagueUsers(t *testing.T) {
 }
 
 func TestUserService_GetOverallRank(t *testing.T) {
-	ctrl := gomock.NewController(t)
+	ctrl, ctx := gomock.WithContext(context.Background(), t)
 	defer ctrl.Finish()
 
-	userClient := usermocks.NewMockUserServiceClient(ctrl)
+	userClient := user_mock.NewMockUserServiceClient(ctrl)
+	tokenGenerator := token_mock.NewMockTokenGenerator(ctrl)
 
-	userService, err := NewUserService(userClient)
+	service, err := New(tokenGenerator, userClient)
 	require.NoError(t, err)
 
 	rank := int64(12345)
 
-	req := &gen.IdRequest{
+	req := &pb.GetOverallRankRequest{
 		Id: id1,
 	}
 
-	resp := &gen.RankResponse{
+	resp := &pb.GetOverallRankResponse{
 		Rank: rank,
 	}
 
+	t.Run("returns error if error generating token", func(t *testing.T) {
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return("", errors.New("error"))
 
-	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
-		userClient.EXPECT().GetOverallRank(context.Background(), req).Return(nil, errors.New("error"))
-
-		result, err := userService.GetOverallRank(id1)
+		result, err := service.GetOverallRank(ctx, id1)
 		require.Error(t, err)
 
 		assert.Empty(t, result)
 	})
 
-	t.Run("gets all users from UserService", func(t *testing.T) {
-		userClient.EXPECT().GetOverallRank(context.Background(), req).Return(resp, nil)
+	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetOverallRank(id1)
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetOverallRank(expectedCtx, req).Return(nil, errors.New("error"))
+
+		result, err := service.GetOverallRank(ctx, id1)
+		require.Error(t, err)
+
+		assert.Empty(t, result)
+	})
+
+	t.Run("gets overall rank from UserService", func(t *testing.T) {
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
+
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetOverallRank(expectedCtx, req).Return(resp, nil)
+
+		result, err := service.GetOverallRank(ctx, id1)
 		require.NoError(t, err)
 
 		assert.Equal(t, rank, result)
@@ -156,41 +204,56 @@ func TestUserService_GetOverallRank(t *testing.T) {
 }
 
 func TestUserService_GetLeagueRank(t *testing.T) {
-	ctrl := gomock.NewController(t)
+	ctrl, ctx := gomock.WithContext(context.Background(), t)
 	defer ctrl.Finish()
 
-	userClient := usermocks.NewMockUserServiceClient(ctrl)
+	userClient := user_mock.NewMockUserServiceClient(ctrl)
+	tokenGenerator := token_mock.NewMockTokenGenerator(ctrl)
 
-	userService, err := NewUserService(userClient)
+	service, err := New(tokenGenerator, userClient)
 	require.NoError(t, err)
 
 	rank := int64(12345)
 
 	ids := []string{id1, id2}
 
-	req := &gen.GroupRankRequest{
-		Id: id1,
+	req := &pb.GetRankForGroupRequest{
+		Id:  id1,
 		Ids: ids,
 	}
 
-	resp := &gen.RankResponse{
+	resp := &pb.GetRankForGroupResponse{
 		Rank: rank,
 	}
 
+	t.Run("returns error if error generating token", func(t *testing.T) {
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return("", errors.New("error"))
+
+		result, err := service.GetLeagueRank(ctx, id1, ids)
+		require.Error(t, err)
+
+		assert.Empty(t, result)
+	})
 
 	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
-		userClient.EXPECT().GetRankForGroup(context.Background(), req).Return(nil, errors.New("error"))
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetLeagueRank(id1, ids)
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetRankForGroup(expectedCtx, req).Return(nil, errors.New("error"))
+
+		result, err := service.GetLeagueRank(ctx, id1, ids)
 		require.Error(t, err)
 
 		assert.Empty(t, result)
 	})
 
 	t.Run("gets all users from UserService", func(t *testing.T) {
-		userClient.EXPECT().GetRankForGroup(context.Background(), req).Return(resp, nil)
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetLeagueRank(id1, ids)
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetRankForGroup(expectedCtx, req).Return(resp, nil)
+
+		result, err := service.GetLeagueRank(ctx, id1, ids)
 		require.NoError(t, err)
 
 		assert.Equal(t, rank, result)
@@ -198,33 +261,49 @@ func TestUserService_GetLeagueRank(t *testing.T) {
 }
 
 func TestUserService_GetUserCount(t *testing.T) {
-	ctrl := gomock.NewController(t)
+	ctrl, ctx := gomock.WithContext(context.Background(), t)
 	defer ctrl.Finish()
 
-	userClient := usermocks.NewMockUserServiceClient(ctrl)
+	userClient := user_mock.NewMockUserServiceClient(ctrl)
+	tokenGenerator := token_mock.NewMockTokenGenerator(ctrl)
 
-	userService, err := NewUserService(userClient)
+	service, err := New(tokenGenerator, userClient)
 	require.NoError(t, err)
 
 	count := int64(1234)
 
-	resp := &gen.CountResponse{
+	resp := &pb.GetUserCountResponse{
 		Count: count,
 	}
 
-	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
-		userClient.EXPECT().GetUserCount(context.Background(), &empty.Empty{}).Return(nil, errors.New("error"))
+	t.Run("returns error if error generating token", func(t *testing.T) {
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return("", errors.New("error"))
 
-		result, err := userService.GetUserCount()
+		result, err := service.GetUserCount(ctx)
+		require.Error(t, err)
+
+		assert.Empty(t, result)
+	})
+
+	t.Run("returns error if there is a problem retrieving users", func(t *testing.T) {
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
+
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetUserCount(expectedCtx, &pb.GetUserCountRequest{}).Return(nil, errors.New("error"))
+
+		result, err := service.GetUserCount(ctx)
 		require.Error(t, err)
 
 		assert.Empty(t, result)
 	})
 
 	t.Run("gets all users from UserService", func(t *testing.T) {
-		userClient.EXPECT().GetUserCount(context.Background(), &empty.Empty{}).Return(resp, nil)
+		expectedCtx := metadata.AppendToOutgoingContext(ctx, "token", token)
 
-		result, err := userService.GetUserCount()
+		tokenGenerator.EXPECT().Generate(ctx, "user").Return(token, nil)
+		userClient.EXPECT().GetUserCount(expectedCtx, &pb.GetUserCountRequest{}).Return(resp, nil)
+
+		result, err := service.GetUserCount(ctx)
 		require.NoError(t, err)
 
 		assert.Equal(t, count, result)
