@@ -1,20 +1,34 @@
 package grpc
 
 import (
+	"context"
 	"errors"
 	"time"
 
-	gen "github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen"
-	common "github.com/cshep4/premier-predictor-microservices/src/common/model"
 	"github.com/cshep4/premier-predictor-microservices/src/livematchservice/internal/handler"
 	"github.com/cshep4/premier-predictor-microservices/src/livematchservice/internal/model"
+
+	pb "github.com/cshep4/premier-predictor-microservices/proto-gen/model/gen"
+	"github.com/cshep4/premier-predictor-microservices/src/common/log"
+	common "github.com/cshep4/premier-predictor-microservices/src/common/model"
 	"github.com/golang/protobuf/ptypes/empty"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
-type server struct {
-	service  handler.Servicer
-	interval time.Duration
+type (
+	server struct {
+		service  handler.Servicer
+		interval time.Duration
+	}
+	observer struct {
+		update func(matchFacts *common.MatchFacts) error
+	}
+)
+
+func (o observer) Update(matchFacts *common.MatchFacts) error {
+	return o.update(matchFacts)
 }
 
 func New(service handler.Servicer, interval time.Duration) (*server, error) {
@@ -32,11 +46,11 @@ func New(service handler.Servicer, interval time.Duration) (*server, error) {
 }
 
 func (s *server) Register(g *grpc.Server) {
-	gen.RegisterLiveMatchServiceServer(g, s)
+	pb.RegisterLiveMatchServiceServer(g, s)
 }
 
-func (s *server) GetUpcomingMatches(_ *empty.Empty, stream gen.LiveMatchService_GetUpcomingMatchesServer) error {
-	matches, err := s.service.GetUpcomingMatches()
+func (s *server) GetUpcomingMatches(_ *empty.Empty, stream pb.LiveMatchService_GetUpcomingMatchesServer) error {
+	matches, err := s.service.GetUpcomingMatches(stream.Context())
 	if err != nil {
 		return err
 	}
@@ -51,7 +65,7 @@ func (s *server) GetUpcomingMatches(_ *empty.Empty, stream gen.LiveMatchService_
 	for {
 		select {
 		case <-ticker.C:
-			matches, err := s.service.GetUpcomingMatches()
+			matches, err := s.service.GetUpcomingMatches(stream.Context())
 			if err != nil {
 				return nil
 			}
@@ -65,37 +79,102 @@ func (s *server) GetUpcomingMatches(_ *empty.Empty, stream gen.LiveMatchService_
 	}
 }
 
-func (s *server) GetMatchSummary(req *gen.PredictionRequest, stream gen.LiveMatchService_GetMatchSummaryServer) error {
-	r := model.PredictionRequest{
-		UserId:  req.UserId,
-		MatchId: req.MatchId,
-	}
-
-	matchSummary, err := s.service.GetMatchSummary(stream.Context(), r)
+func (s *server) GetMatchSummary(req *pb.PredictionRequest, stream pb.LiveMatchService_GetMatchSummaryServer) error {
+	matchSummary, err := s.service.GetMatchSummary(stream.Context(), model.PredictionRequest{
+		UserId:  req.GetUserId(),
+		MatchId: req.GetMatchId(),
+	})
 	if err != nil {
-		return err
+		log.Error(stream.Context(), "error_getting_match_summary",
+			log.ErrorParam(err),
+			log.SafeParam("matchId", req.GetMatchId()),
+			log.SafeParam("userId", req.GetUserId()),
+		)
+		return status.Error(codes.Internal, "could not get match summary")
+	}
+	res := model.MatchSummaryToGrpc(matchSummary)
+
+	if err := stream.Send(res); err != nil {
+		log.Error(stream.Context(), "error_sending_response",
+			log.ErrorParam(err),
+			log.SafeParam("matchId", req.GetMatchId()),
+			log.SafeParam("userId", req.GetUserId()),
+		)
+		return status.Error(codes.Internal, "could not send response")
 	}
 
-	resp := model.MatchSummaryToGrpc(matchSummary)
+	obvs := observer{update: func(matchFacts *common.MatchFacts) error {
+		res.Match = common.MatchFactsToGrpc(matchFacts)
 
-	if err := stream.Send(resp); err != nil {
-		return err
-	}
-
-	ticker := time.NewTicker(s.interval)
-	for {
-		select {
-		case <-ticker.C:
-			match, err := s.service.GetMatchFacts(req.MatchId)
-			if err != nil {
-				return nil
-			}
-
-			resp.Match = common.MatchFactsToGrpc(match)
-
-			if err := stream.Send(resp); err != nil {
-				return nil
-			}
+		if err := stream.Send(res); err != nil {
+			log.Error(stream.Context(), "error_sending_response",
+				log.ErrorParam(err),
+				log.SafeParam("matchId", req.GetMatchId()),
+				log.SafeParam("userId", req.GetUserId()),
+			)
+			return status.Error(codes.Internal, "could not send response")
 		}
+
+		return nil
+	}}
+
+	err = s.service.SubscribeToMatch(stream.Context(), req.GetMatchId(), obvs)
+	if err != nil {
+		log.Error(stream.Context(), "error_subscribing_to_live_match",
+			log.ErrorParam(err),
+			log.SafeParam("matchId", req.GetMatchId()),
+			log.SafeParam("userId", req.GetUserId()),
+		)
+		return status.Error(codes.Internal, "could not subscribe to live match")
 	}
+
+	return nil
+}
+
+func (s *server) GetTodaysLiveMatches(_ *pb.GetTodaysLiveMatchesRequest, stream pb.LiveMatchService_GetTodaysLiveMatchesServer) error {
+	obvs := observer{update: func(matchFacts *common.MatchFacts) error {
+		if err := stream.Send(&pb.GetTodaysLiveMatchesResponse{Match: common.MatchFactsToGrpc(matchFacts)}); err != nil {
+			log.Error(stream.Context(), "error_sending_response", log.ErrorParam(err))
+			return status.Error(codes.Internal, "could not send response")
+		}
+
+		return nil
+	}}
+
+	err := s.service.SubscribeToTodaysMatches(stream.Context(), obvs)
+	if err != nil {
+		log.Error(stream.Context(), "error_subscribing_to_todays_matches", log.ErrorParam(err))
+		return status.Error(codes.Internal, "could not subscribe to todays matches")
+	}
+
+	return nil
+}
+
+func (s *server) GetLiveMatch(ctx context.Context, req *pb.GetLiveMatchRequest) (*pb.GetLiveMatchResponse, error) {
+	match, err := s.service.GetMatchFacts(ctx, req.GetId())
+	if err != nil {
+		log.Error(ctx, "error_getting_live_match", log.ErrorParam(err))
+		return nil, status.Error(codes.Internal, "could not get live match")
+	}
+
+	return &pb.GetLiveMatchResponse{
+		Match: common.MatchFactsToGrpc(match),
+	}, nil
+}
+
+func (s *server) ListTodaysMatches(ctx context.Context, _ *pb.ListTodaysMatchesRequest) (*pb.ListTodaysMatchesResponse, error) {
+	matches, err := s.service.GetTodaysMatches(ctx)
+	if err != nil {
+		log.Error(ctx, "error_getting_todays_matches", log.ErrorParam(err))
+		return nil, status.Error(codes.Internal, "could not list today's matches")
+	}
+
+	var matchFacts []*pb.MatchFacts
+	for _, m := range matches {
+		matchFacts = append(matchFacts, common.MatchFactsToGrpc(&m))
+	}
+
+	return &pb.ListTodaysMatchesResponse{
+		Matches: matchFacts,
+	}, nil
 }
