@@ -7,10 +7,32 @@ import (
 	"time"
 )
 
-type MatchSummary struct {
-	Match             *model.MatchFacts             `json:"match"`
-	PredictionSummary *model.MatchPredictionSummary `json:"predictionSummary"`
-	Prediction        *model.Prediction             `json:"prediction"`
+type (
+	MatchObserver interface {
+		Update(matchFacts *model.MatchFacts) error
+	}
+	MatchObservable struct {
+		observers []MatchObserver
+	}
+	MatchSummary struct {
+		Match             *model.MatchFacts             `json:"match"`
+		PredictionSummary *model.MatchPredictionSummary `json:"predictionSummary"`
+		Prediction        *model.Prediction             `json:"prediction"`
+	}
+)
+
+func (m *MatchObservable) AddObserver(observer MatchObserver) {
+	m.observers = append(m.observers, observer)
+}
+
+func (m *MatchObservable) Notify(matchFacts *model.MatchFacts) error {
+	for _, o := range m.observers {
+		if err := o.Update(matchFacts); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 func MatchSummaryFromGrpc(matchFacts *gen.MatchSummary) *MatchSummary {
