@@ -3,6 +3,8 @@ package mongo
 import (
 	"context"
 	"errors"
+	"fmt"
+	"golang.org/x/sync/errgroup"
 	"time"
 
 	common "github.com/cshep4/premier-predictor-microservices/src/common/model"
@@ -15,7 +17,7 @@ import (
 
 const (
 	db         = "liveMatch"
-	collection = "liveMatch"
+	collection = "2020-2021"
 )
 
 var limit = int64(20)
@@ -26,7 +28,7 @@ type store struct {
 
 func New(ctx context.Context, client *mongo.Client) (*store, error) {
 	if client == nil {
-		return nil, errors.New("mongo_client_is_nil")
+		return nil, errors.New("mongo client is nil")
 	}
 
 	s := &store{
@@ -44,8 +46,8 @@ func New(ctx context.Context, client *mongo.Client) (*store, error) {
 	return s, nil
 }
 
-func (r *store) ensureIndexes(ctx context.Context) error {
-	_, err := r.client.
+func (s *store) ensureIndexes(ctx context.Context) error {
+	_, err := s.client.
 		Database(db).
 		Collection(collection).
 		Indexes().CreateOne(
@@ -69,19 +71,40 @@ func (r *store) ensureIndexes(ctx context.Context) error {
 	return nil
 }
 
-func (r *store) GetUpcomingMatches() ([]common.MatchFacts, error) {
+func (s *store) GetUpcomingMatches(ctx context.Context) ([]common.MatchFacts, error) {
 	year, month, day := time.Now().Date()
 	today := time.Date(year, month, day, 0, 0, 0, 0, time.Now().Location())
 
-	return r.getMatches(
+	return s.getMatches(
+		ctx,
+		bson.D{
+			{
+				Key:   "matchDate",
+				Value: bson.D{{Key: "$gte", Value: today}},
+			},
+		},
+		&options.FindOptions{
+			Limit: &limit,
+			Sort: bson.D{
+				bson.E{Key: "matchDate", Value: 1},
+			},
+		},
+	)
+}
+
+func (s *store) GetTodaysMatches(ctx context.Context) ([]common.MatchFacts, error) {
+	year, month, day := time.Now().Date()
+	today := time.Date(year, month, day-1, 0, 0, 0, 0, time.Now().Location())
+	tomorrow := time.Date(year, month, day+1, 0, 0, 0, 0, time.Now().Location())
+
+	return s.getMatches(
+		ctx,
 		bson.D{
 			{
 				Key: "matchDate",
 				Value: bson.D{
-					{
-						Key:   "$gte",
-						Value: today,
-					},
+					{Key: "$gte", Value: today},
+					{Key: "$lt", Value: tomorrow},
 				},
 			},
 		},
@@ -94,10 +117,8 @@ func (r *store) GetUpcomingMatches() ([]common.MatchFacts, error) {
 	)
 }
 
-func (r *store) getMatches(filter interface{}, opts *options.FindOptions) ([]common.MatchFacts, error) {
-	ctx := context.Background()
-
-	cur, err := r.client.
+func (s *store) getMatches(ctx context.Context, filter interface{}, opts *options.FindOptions) ([]common.MatchFacts, error) {
+	cur, err := s.client.
 		Database(db).
 		Collection(collection).
 		Find(
@@ -130,17 +151,15 @@ func (r *store) getMatches(filter interface{}, opts *options.FindOptions) ([]com
 	return matches, nil
 }
 
-func (r *store) GetMatchFacts(id string) (*common.MatchFacts, error) {
+func (s *store) GetMatchFacts(ctx context.Context, id string) (*common.MatchFacts, error) {
 	var m matchFactsEntity
 
-	err := r.client.
+	err := s.client.
 		Database(db).
 		Collection(collection).
 		FindOne(
-			context.Background(),
-			bson.M{
-				"_id": id,
-			},
+			ctx,
+			bson.M{"_id": id},
 		).
 		Decode(&m)
 
@@ -153,6 +172,110 @@ func (r *store) GetMatchFacts(id string) (*common.MatchFacts, error) {
 	}
 
 	return toMatchFacts(&m), nil
+}
+
+func (s *store) SubscribeToMatch(ctx context.Context, id string, observer model.MatchObserver) error {
+	cs, err := s.client.
+		Database(db).
+		Collection(collection).
+		Watch(ctx, mongo.Pipeline{
+			bson.D{{
+				Key: "$match",
+				Value: bson.D{
+					{Key: "$and", Value: bson.A{
+						bson.D{{Key: "$or", Value: bson.A{
+							bson.D{{Key: "operationType", Value: "replace"}},
+							bson.D{{Key: "operationType", Value: "update"}},
+						}}},
+						bson.D{{Key: "documentKey._id", Value: id}},
+					}},
+				},
+			}},
+		})
+	if err != nil {
+		return fmt.Errorf("watch_match: %w", err)
+	}
+
+	observable := model.MatchObservable{}
+	observable.AddObserver(observer)
+
+	g, ctx := errgroup.WithContext(ctx)
+
+	g.Go(func() error {
+		for cs.Next(ctx) {
+			m, err := s.GetMatchFacts(ctx, id)
+			if err != nil {
+				return fmt.Errorf("get_match_facts: %w", err)
+			}
+
+			if err := observable.Notify(m); err != nil {
+				return fmt.Errorf("observable_notify: %w", err)
+			}
+		}
+		return nil
+	})
+
+	return g.Wait()
+}
+
+func (s *store) SubscribeToMatches(ctx context.Context, ids []string, observer model.MatchObserver) error {
+	cs, err := s.client.
+		Database(db).
+		Collection(collection).
+		Watch(ctx, mongo.Pipeline{
+			bson.D{{
+				Key: "$match",
+				Value: bson.D{
+					{Key: "$and", Value: bson.A{
+						bson.D{{Key: "$or", Value: bson.A{
+							bson.D{{Key: "operationType", Value: "replace"}},
+							bson.D{{Key: "operationType", Value: "update"}},
+						}}},
+						bson.D{{Key: "documentKey._id", Value: bson.D{{Key: "$in", Value: ids}}}},
+					}},
+				},
+			}},
+		})
+	if err != nil {
+		return fmt.Errorf("watch_matches: %w", err)
+	}
+
+	observable := model.MatchObservable{}
+	observable.AddObserver(observer)
+
+	g, ctx := errgroup.WithContext(ctx)
+
+	g.Go(func() error {
+		for cs.Next(ctx) {
+			var r map[string]interface{}
+			err := cs.Decode(&r)
+			if err != nil {
+				return fmt.Errorf("cs_decode: %w", err)
+			}
+
+			dd, ok := r["documentKey"].(map[string]interface{})
+			if !ok {
+				return fmt.Errorf("invalid documentKey: %v", dd)
+			}
+
+			id, ok := dd["_id"].(string)
+			if !ok {
+				return fmt.Errorf("invalid _id: %v", id)
+			}
+
+			m, err := s.GetMatchFacts(ctx, id)
+			if err != nil {
+				return fmt.Errorf("get_match_facts: %w", err)
+			}
+
+			if err := observable.Notify(m); err != nil {
+				return fmt.Errorf("observable_notify: %w", err)
+			}
+		}
+		return nil
+	})
+
+	return g.Wait()
 }
 
 func (s *store) Ping(ctx context.Context) error {
