@@ -15,7 +15,6 @@ import (
 	"github.com/cshep4/premier-predictor-microservices/src/common/gcp/tracer"
 	grpcconn "github.com/cshep4/premier-predictor-microservices/src/common/grpc"
 	"github.com/cshep4/premier-predictor-microservices/src/common/log"
-	"github.com/cshep4/premier-predictor-microservices/src/common/run"
 	"github.com/cshep4/premier-predictor-microservices/src/common/runner/grpc"
 	"github.com/cshep4/premier-predictor-microservices/src/common/runner/http"
 	"github.com/cshep4/premier-predictor-microservices/src/common/store/mongo"
@@ -24,7 +23,6 @@ import (
 	"github.com/cshep4/premier-predictor-microservices/src/livematchservice/internal/prediction"
 	svc "github.com/cshep4/premier-predictor-microservices/src/livematchservice/internal/service"
 	mongostore "github.com/cshep4/premier-predictor-microservices/src/livematchservice/internal/store/mongo"
-	"golang.org/x/sync/errgroup"
 )
 
 const (
@@ -76,11 +74,6 @@ func start(ctx context.Context) error {
 	defer predictionConn.Close()
 	predictionClient := gen.NewPredictionServiceClient(predictionConn)
 
-	authenticator, err := auth.New(authClient)
-	if err != nil {
-		return fmt.Errorf("create_authenticator: %w", err)
-	}
-
 	predictor, err := prediction.New(predictionClient)
 	if err != nil {
 		return fmt.Errorf("create_predictor: %w", err)
@@ -112,6 +105,11 @@ func start(ctx context.Context) error {
 		return fmt.Errorf("create_grpc_handler: %w", err)
 	}
 
+	authenticator, err := auth.New(authClient, "livematch", h)
+	if err != nil {
+		return fmt.Errorf("create_authenticator: %w", err)
+	}
+
 	tracer := tracer.New()
 
 	app := app.New(
@@ -124,8 +122,8 @@ func start(ctx context.Context) error {
 			grpc.New(
 				grpc.WithPort(grpcPort),
 				grpc.WithLogger(serviceName, logLevel),
-				grpc.WithUnaryInterceptor(tracer.GrpcUnary),
-				grpc.WithStreamInterceptor(tracer.GrpcStream),
+				//grpc.WithUnaryInterceptor(tracer.GrpcUnary),
+				//grpc.WithStreamInterceptor(tracer.GrpcStream),
 				grpc.WithUnaryInterceptor(authenticator.GrpcUnary),
 				grpc.WithStreamInterceptor(authenticator.GrpcStream),
 				grpc.WithRegisterer(rpc),
@@ -143,13 +141,7 @@ func start(ctx context.Context) error {
 		),
 	)
 
-	ctx, cancel := context.WithCancel(ctx)
-	g, ctx := errgroup.WithContext(ctx)
-
-	g.Go(func() error { return app.Run(ctx) })
-	g.Go(run.HandleShutdown(g, ctx, cancel, app.Shutdown))
-
-	return g.Wait()
+	return app.Run(ctx)
 }
 
 func main() {
