@@ -1,16 +1,17 @@
 package user
 
 import (
-	common "github.com/cshep4/premier-predictor-microservices/src/common/model"
-	"github.com/cshep4/premier-predictor-microservices/src/common/util"
+	"context"
+	"errors"
+	"testing"
+
+	"github.com/cshep4/premier-predictor-microservices/src/userservice/internal/mocks/store"
 	"github.com/cshep4/premier-predictor-microservices/src/userservice/internal/model"
-	"github.com/cshep4/premier-predictor-microservices/src/userservice/internal/repository/mocks"
+
 	"github.com/golang/mock/gomock"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/crypto/bcrypt"
-	"testing"
 )
 
 const (
@@ -22,36 +23,36 @@ const (
 )
 
 var (
-	e = errors.New("error")
+	testErr = errors.New("error")
 )
 
 func TestService_GetUserById(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
-	t.Run("Returns error if there is a problem", func(t *testing.T) {
-		repository.EXPECT().GetUserById(userId).Return(nil, e)
+	t.Run("returns error if error getting user", func(t *testing.T) {
+		store.EXPECT().GetUserById(context.Background(), userId).Return(nil, testErr)
 
-		result, err := service.GetUserById(userId)
-
+		result, err := service.GetUserById(context.Background(), userId)
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+
+		assert.True(t, errors.Is(err, testErr))
 		assert.Nil(t, result)
 	})
 
-	t.Run("Gets user from db", func(t *testing.T) {
+	t.Run("gets user from db", func(t *testing.T) {
 		user := &model.User{
 			Id: userId,
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
-		result, err := service.GetUserById(userId)
+		result, err := service.GetUserById(context.Background(), userId)
 
 		require.NoError(t, err)
 		assert.Equal(t, user, result)
@@ -62,9 +63,9 @@ func TestService_UpdateUserInfo(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
 	t.Run("returns error if the email address is not valid", func(t *testing.T) {
@@ -72,11 +73,13 @@ func TestService_UpdateUserInfo(t *testing.T) {
 			Email: "invalid email address",
 		}
 
-		err := service.UpdateUserInfo(userInfo)
+		err := service.UpdateUserInfo(context.Background(), userInfo)
 		require.Error(t, err)
 
-		assert.Equal(t, invalidEmail, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "email", ipe.Parameter)
 	})
 
 	t.Run("returns error if the email address is already taken by a different user", func(t *testing.T) {
@@ -85,13 +88,15 @@ func TestService_UpdateUserInfo(t *testing.T) {
 			Email: emailAddress,
 		}
 
-		repository.EXPECT().IsEmailTakenByADifferentUser(userId, emailAddress).Return(true)
+		store.EXPECT().IsEmailTakenByADifferentUser(context.Background(), userId, emailAddress).Return(true)
 
-		err := service.UpdateUserInfo(userInfo)
+		err := service.UpdateUserInfo(context.Background(), userInfo)
 		require.Error(t, err)
 
-		assert.Equal(t, emailAlreadyTaken, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "email already taken", ipe.Parameter)
 	})
 
 	t.Run("returns error if the first name is blank", func(t *testing.T) {
@@ -100,13 +105,15 @@ func TestService_UpdateUserInfo(t *testing.T) {
 			Email: emailAddress,
 		}
 
-		repository.EXPECT().IsEmailTakenByADifferentUser(userId, emailAddress).Return(false)
+		store.EXPECT().IsEmailTakenByADifferentUser(context.Background(), userId, emailAddress).Return(false)
 
-		err := service.UpdateUserInfo(userInfo)
+		err := service.UpdateUserInfo(context.Background(), userInfo)
 		require.Error(t, err)
 
-		assert.Equal(t, firstNameIsBlank, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "first name", ipe.Parameter)
 	})
 
 	t.Run("returns error if the surname is blank", func(t *testing.T) {
@@ -116,13 +123,15 @@ func TestService_UpdateUserInfo(t *testing.T) {
 			FirstName: "first name",
 		}
 
-		repository.EXPECT().IsEmailTakenByADifferentUser(userId, emailAddress).Return(false)
+		store.EXPECT().IsEmailTakenByADifferentUser(context.Background(), userId, emailAddress).Return(false)
 
-		err := service.UpdateUserInfo(userInfo)
+		err := service.UpdateUserInfo(context.Background(), userInfo)
 		require.Error(t, err)
 
-		assert.Equal(t, surnameIsBlank, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "surname", ipe.Parameter)
 	})
 
 	t.Run("returns error if details cannot be updated", func(t *testing.T) {
@@ -133,13 +142,13 @@ func TestService_UpdateUserInfo(t *testing.T) {
 			Surname:   "surname",
 		}
 
-		repository.EXPECT().IsEmailTakenByADifferentUser(userId, emailAddress).Return(false)
-		repository.EXPECT().UpdateUserInfo(userInfo).Return(e)
+		store.EXPECT().IsEmailTakenByADifferentUser(context.Background(), userId, emailAddress).Return(false)
+		store.EXPECT().UpdateUserInfo(context.Background(), userInfo).Return(testErr)
 
-		err := service.UpdateUserInfo(userInfo)
+		err := service.UpdateUserInfo(context.Background(), userInfo)
 		require.Error(t, err)
 
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 	})
 
 	t.Run("returns nil if details are updated successfully", func(t *testing.T) {
@@ -150,10 +159,10 @@ func TestService_UpdateUserInfo(t *testing.T) {
 			Surname:   "surname",
 		}
 
-		repository.EXPECT().IsEmailTakenByADifferentUser(userId, emailAddress).Return(false)
-		repository.EXPECT().UpdateUserInfo(userInfo).Return(nil)
+		store.EXPECT().IsEmailTakenByADifferentUser(context.Background(), userId, emailAddress).Return(false)
+		store.EXPECT().UpdateUserInfo(context.Background(), userInfo).Return(nil)
 
-		err := service.UpdateUserInfo(userInfo)
+		err := service.UpdateUserInfo(context.Background(), userInfo)
 		require.NoError(t, err)
 
 		assert.Nil(t, err)
@@ -164,20 +173,18 @@ func TestService_UpdatePassword(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
-	//newHashedPassword, _ := bcrypt.GenerateFromPassword([]byte(newValidPassword), 10)
-
 	t.Run("returns error if the user cannot be retrieved", func(t *testing.T) {
-		repository.EXPECT().GetUserById(userId).Return(nil, e)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(nil, testErr)
 
-		err := service.UpdatePassword(model.UpdatePassword{Id: userId})
+		err := service.UpdateUserPassword(context.Background(), model.UpdatePassword{Id: userId})
 		require.Error(t, err)
 
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 	})
 
 	t.Run("returns error if old password does not match what is currently stored", func(t *testing.T) {
@@ -186,18 +193,20 @@ func TestService_UpdatePassword(t *testing.T) {
 			Password: hashPassword(oldPassword),
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
 		updatePassword := model.UpdatePassword{
 			Id:          userId,
 			OldPassword: "different old password",
 		}
 
-		err := service.UpdatePassword(updatePassword)
+		err := service.UpdateUserPassword(context.Background(), updatePassword)
 		require.Error(t, err)
 
-		assert.Equal(t, oldPasswordDoesNotMatch, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "old password does not match", ipe.Parameter)
 	})
 
 	t.Run("returns error if new password does not match the confirmation", func(t *testing.T) {
@@ -206,7 +215,7 @@ func TestService_UpdatePassword(t *testing.T) {
 			Password: hashPassword(oldPassword),
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
 		updatePassword := model.UpdatePassword{
 			Id:              userId,
@@ -215,11 +224,13 @@ func TestService_UpdatePassword(t *testing.T) {
 			ConfirmPassword: "different confirmation password",
 		}
 
-		err := service.UpdatePassword(updatePassword)
+		err := service.UpdateUserPassword(context.Background(), updatePassword)
 		require.Error(t, err)
 
-		assert.Equal(t, confirmationDoesNotMatch, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "confirmation does not match", ipe.Parameter)
 	})
 
 	t.Run("returns error if new password is not valid", func(t *testing.T) {
@@ -228,7 +239,7 @@ func TestService_UpdatePassword(t *testing.T) {
 			Password: hashPassword(oldPassword),
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
 		updatePassword := model.UpdatePassword{
 			Id:              userId,
@@ -237,11 +248,13 @@ func TestService_UpdatePassword(t *testing.T) {
 			ConfirmPassword: newPassword,
 		}
 
-		err := service.UpdatePassword(updatePassword)
+		err := service.UpdateUserPassword(context.Background(), updatePassword)
 		require.Error(t, err)
 
-		assert.Equal(t, invalidPassword, util.GetErrorMessage(err))
-		assert.Equal(t, common.ErrInvalidRequestData, errors.Cause(err))
+		ipe, ok := err.(model.InvalidParameterError)
+		require.True(t, ok)
+
+		assert.Equal(t, "new password", ipe.Parameter)
 	})
 
 	t.Run("returns error if password cannot be updated", func(t *testing.T) {
@@ -250,7 +263,7 @@ func TestService_UpdatePassword(t *testing.T) {
 			Password: hashPassword(oldPassword),
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
 		updatePassword := model.UpdatePassword{
 			Id:              userId,
@@ -259,12 +272,12 @@ func TestService_UpdatePassword(t *testing.T) {
 			ConfirmPassword: newValidPassword,
 		}
 
-		repository.EXPECT().UpdatePassword(userId, gomock.Any()).Return(e)
+		store.EXPECT().UpdatePassword(context.Background(), userId, gomock.Any()).Return(testErr)
 
-		err := service.UpdatePassword(updatePassword)
+		err := service.UpdateUserPassword(context.Background(), updatePassword)
 		require.Error(t, err)
 
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 	})
 
 	t.Run("returns nil if password is updated successfully", func(t *testing.T) {
@@ -273,7 +286,7 @@ func TestService_UpdatePassword(t *testing.T) {
 			Password: hashPassword(oldPassword),
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
 		updatePassword := model.UpdatePassword{
 			Id:              userId,
@@ -282,9 +295,9 @@ func TestService_UpdatePassword(t *testing.T) {
 			ConfirmPassword: newValidPassword,
 		}
 
-		repository.EXPECT().UpdatePassword(userId, gomock.Any()).Return(nil)
+		store.EXPECT().UpdatePassword(context.Background(), userId, gomock.Any()).Return(nil)
 
-		err := service.UpdatePassword(updatePassword)
+		err := service.UpdateUserPassword(context.Background(), updatePassword)
 		require.NoError(t, err)
 
 		assert.Nil(t, err)
@@ -295,18 +308,18 @@ func TestService_GetUserScore(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
 	t.Run("returns error if user cannot be retrieved", func(t *testing.T) {
-		repository.EXPECT().GetUserById(userId).Return(nil, e)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(nil, testErr)
 
-		result, err := service.GetUserScore(userId)
+		result, err := service.GetUserScore(context.Background(), userId)
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Equal(t, 0, result)
 	})
 
@@ -318,9 +331,9 @@ func TestService_GetUserScore(t *testing.T) {
 			Score: score,
 		}
 
-		repository.EXPECT().GetUserById(userId).Return(user, nil)
+		store.EXPECT().GetUserById(context.Background(), userId).Return(user, nil)
 
-		result, err := service.GetUserScore(userId)
+		result, err := service.GetUserScore(context.Background(), userId)
 
 		require.NoError(t, err)
 		assert.Equal(t, score, result)
@@ -336,18 +349,18 @@ func TestService_GetAllUsers(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
 	t.Run("returns error if users cannot be retrieved", func(t *testing.T) {
-		repository.EXPECT().GetAllUsers().Return(nil, e)
+		store.EXPECT().GetAllUsers(context.Background()).Return(nil, testErr)
 
-		result, err := service.GetAllUsers()
+		result, err := service.GetAllUsers(context.Background())
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Nil(t, result)
 	})
 
@@ -358,9 +371,9 @@ func TestService_GetAllUsers(t *testing.T) {
 			},
 		}
 
-		repository.EXPECT().GetAllUsers().Return(users, nil)
+		store.EXPECT().GetAllUsers(context.Background()).Return(users, nil)
 
-		result, err := service.GetAllUsers()
+		result, err := service.GetAllUsers(context.Background())
 
 		require.NoError(t, err)
 
@@ -372,20 +385,20 @@ func TestService_GetAllUsersByIds(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
 	ids := []string{userId}
 
 	t.Run("returns error if users cannot be retrieved", func(t *testing.T) {
-		repository.EXPECT().GetAllUsersByIds(ids).Return(nil, e)
+		store.EXPECT().GetAllUsersByIds(context.Background(), ids).Return(nil, testErr)
 
-		result, err := service.GetAllUsersByIds(ids)
+		result, err := service.GetAllUsersByIds(context.Background(), ids)
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Nil(t, result)
 	})
 
@@ -396,9 +409,9 @@ func TestService_GetAllUsersByIds(t *testing.T) {
 			},
 		}
 
-		repository.EXPECT().GetAllUsersByIds(ids).Return(users, nil)
+		store.EXPECT().GetAllUsersByIds(context.Background(), ids).Return(users, nil)
 
-		result, err := service.GetAllUsersByIds(ids)
+		result, err := service.GetAllUsersByIds(context.Background(), ids)
 
 		require.NoError(t, err)
 
@@ -410,27 +423,27 @@ func TestService_GetRankForGroup(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
 	ids := []string{userId}
 
-	t.Run("Returns error if there is a problem", func(t *testing.T) {
-		repository.EXPECT().GetRankForGroup(userId, ids).Return(int64(0), e)
+	t.Run("returns error if error getting rank", func(t *testing.T) {
+		store.EXPECT().GetRankForGroup(context.Background(), userId, ids).Return(int64(0), testErr)
 
-		result, err := service.GetRankForGroup(userId, ids)
+		result, err := service.GetRankForGroup(context.Background(), userId, ids)
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Empty(t, result)
 	})
 
-	t.Run("Gets rank for group", func(t *testing.T) {
-		repository.EXPECT().GetRankForGroup(userId, ids).Return(int64(1), nil)
+	t.Run("gets rank for group", func(t *testing.T) {
+		store.EXPECT().GetRankForGroup(context.Background(), userId, ids).Return(int64(1), nil)
 
-		result, err := service.GetRankForGroup(userId, ids)
+		result, err := service.GetRankForGroup(context.Background(), userId, ids)
 
 		require.NoError(t, err)
 		assert.Equal(t, int64(1), result)
@@ -441,25 +454,25 @@ func TestService_GetOverallRank(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
-	t.Run("Returns error if there is a problem", func(t *testing.T) {
-		repository.EXPECT().GetOverallRank(userId).Return(int64(0), e)
+	t.Run("returns error if there is a problem", func(t *testing.T) {
+		store.EXPECT().GetOverallRank(context.Background(), userId).Return(int64(0), testErr)
 
-		result, err := service.GetOverallRank(userId)
+		result, err := service.GetOverallRank(context.Background(), userId)
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Empty(t, result)
 	})
 
-	t.Run("Gets overall rank", func(t *testing.T) {
-		repository.EXPECT().GetOverallRank(userId).Return(int64(1), nil)
+	t.Run("gets overall rank", func(t *testing.T) {
+		store.EXPECT().GetOverallRank(context.Background(), userId).Return(int64(1), nil)
 
-		result, err := service.GetOverallRank(userId)
+		result, err := service.GetOverallRank(context.Background(), userId)
 
 		require.NoError(t, err)
 		assert.Equal(t, int64(1), result)
@@ -470,25 +483,25 @@ func TestService_GetUserCount(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
-	t.Run("Returns error if there is a problem", func(t *testing.T) {
-		repository.EXPECT().GetUserCount().Return(int64(0), e)
+	t.Run("returns error if there is a problem", func(t *testing.T) {
+		store.EXPECT().GetUserCount(context.Background()).Return(int64(0), testErr)
 
-		result, err := service.GetUserCount()
+		result, err := service.GetUserCount(context.Background())
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Empty(t, result)
 	})
 
-	t.Run("Gets user count", func(t *testing.T) {
-		repository.EXPECT().GetUserCount().Return(int64(1), nil)
+	t.Run("gets user count", func(t *testing.T) {
+		store.EXPECT().GetUserCount(context.Background()).Return(int64(1), nil)
 
-		result, err := service.GetUserCount()
+		result, err := service.GetUserCount(context.Background())
 
 		require.NoError(t, err)
 		assert.Equal(t, int64(1), result)
@@ -499,30 +512,30 @@ func TestService_GetUserByEmail(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	repository := usermocks.NewMockRepository(ctrl)
+	store := store_mock.NewMockStore(ctrl)
 
-	service, err := NewService(repository)
+	service, err := New(store)
 	require.NoError(t, err)
 
-	t.Run("Returns error if there is a problem", func(t *testing.T) {
-		repository.EXPECT().GetUserByEmail(emailAddress).Return(nil, e)
+	t.Run("returns error if there is a problem", func(t *testing.T) {
+		store.EXPECT().GetUserByEmail(context.Background(), emailAddress).Return(nil, testErr)
 
-		result, err := service.GetUserByEmail(emailAddress)
+		result, err := service.GetUserByEmail(context.Background(), emailAddress)
 
 		require.Error(t, err)
-		assert.Equal(t, e, err)
+		assert.True(t, errors.Is(err, testErr))
 		assert.Nil(t, result)
 	})
 
-	t.Run("Gets user from db", func(t *testing.T) {
+	t.Run("gets user from db", func(t *testing.T) {
 		user := &model.User{
 			Id:    userId,
 			Email: emailAddress,
 		}
 
-		repository.EXPECT().GetUserByEmail(emailAddress).Return(user, nil)
+		store.EXPECT().GetUserByEmail(context.Background(), emailAddress).Return(user, nil)
 
-		result, err := service.GetUserByEmail(emailAddress)
+		result, err := service.GetUserByEmail(context.Background(), emailAddress)
 
 		require.NoError(t, err)
 		assert.Equal(t, user, result)
